Split core/main_test.ts
