+++ cli/index_project_test.ts
@@ -4,10 +4,9 @@ import * as path from "path";
 
 import {
   runCli,
-  setupProject,
-  writeDefinitionFile
+  setupProject
 } from "df/cli/index_test_base";
-import { suite, test } from "df/testing";
+import { suite, test, writeDefinitionFile } from "df/testing";
 import { TmpDirFixture } from "df/testing/fixtures";
 
 suite("project ops", ({ afterEach }) => {
