+++ cli/tests/jit/index_jit_advanced_test.ts
@@ -3,10 +3,9 @@ import { expect } from "chai";
 import {
   CREDENTIALS_PATH,
   runCli,
-  setupJitProject,
-  writeDefinitionFile
+  setupJitProject
 } from "df/cli/index_test_base";
-import { suite, test } from "df/testing";
+import { suite, test, writeDefinitionFile } from "df/testing";
 import { TmpDirFixture } from "df/testing/fixtures";
 
 suite("JiT support advanced", ({ afterEach }) => {
