+++ cli/tests/jit/index_jit_runtime_test.ts
@@ -5,10 +5,9 @@ import * as path from "path";
 import {
   CREDENTIALS_PATH,
   runCli,
-  setupJitProject,
-  writeDefinitionFile
+  setupJitProject
 } from "df/cli/index_test_base";
-import { suite, test } from "df/testing";
+import { suite, test, writeDefinitionFile } from "df/testing";
 import { TmpDirFixture } from "df/testing/fixtures";
 
 suite("JiT support runtime", ({ afterEach }) => {
