+++ core/BUILD
@@ -81,7 +81,13 @@ ts_test_suite(
         "compilers_test.ts",
         "jit_compiler_test.ts",
         "jit_context_test.ts",
-        "main_test.ts",
+        "main_action_configs_test.ts",
+        "main_extensions_test.ts",
+        "main_js_api_test.ts",
+        "main_property_graphs_test.ts",
+        "main_session_test.ts",
+        "main_sqlx_chars_test.ts",
+        "main_workflow_settings_test.ts",
         "utils_test.ts",
         "workflow_settings_test.ts",
     ],
