--- core/main_test.ts
@@ -1,4215 +1,0 @@
-import { expect } from "chai";
-import * as fs from "fs-extra";
-import { dump as dumpYaml } from "js-yaml";
-import * as path from "path";
-
-import { version } from "df/core/version";
-import { dataform, google } from "df/protos/ts";
-import { asPlainObject, suite, test } from "df/testing";
-import { TmpDirFixture } from "df/testing/fixtures";
-import {
-  coreExecutionRequestFromPath,
-  runMainInVm,
-  VALID_DATAFORM_JSON,
-  VALID_WORKFLOW_SETTINGS_YAML,
-  WorkflowSettingsTemplates
-} from "df/testing/run_core";
-
-const EMPTY_NOTEBOOK_CONTENTS = '{ "cells": [] }';
-
-interface IVerifiableAction {
-  type?: string | null,
-  target?: dataform.ITarget | null;
-  canonicalTarget?: dataform.ITarget | null;
-  dependencyTargets?: dataform.ITarget[] | null
-}
-
-function toVerifiableAction(graph: dataform.ICompiledGraph, actionType: string): IVerifiableAction {
-  let action: dataform.IAssertion | dataform.ITable | dataform.IPropertyGraph
-  switch (actionType) {
-    case "assertion":
-      action = graph.assertions[0];
-      break;
-    case "operations":
-      action = graph.operations[0];
-      break;
-    case "propertyGraph":
-      action = graph.propertyGraphs[0];
-      break;
-    default:
-      action = graph.tables[0];
-  }
-
-  return {
-    type: actionType,
-    target: action.target,
-    canonicalTarget: action.canonicalTarget,
-    dependencyTargets: action.dependencyTargets,
-  };
-}
-
-// INFO: if you want to see an overview of the tests in this file, press cmd-k-3 while in
-// VSCode, to collapse everything below the third level of indentation.
-suite("@dataform/core", ({ afterEach }) => {
-  const tmpDirFixture = new TmpDirFixture(afterEach);
-
-  suite("session", () => {
-    suite("resolve succeeds", () => {
-      [
-        WorkflowSettingsTemplates.bigquery,
-        WorkflowSettingsTemplates.bigqueryWithDatasetSuffix,
-        WorkflowSettingsTemplates.bigqueryWithNamePrefix
-      ].forEach(testConfig => {
-        test(`resolve with name prefix "${testConfig.namePrefix}" and dataset suffix "${testConfig.datasetSuffix}"`, () => {
-          const projectDir = tmpDirFixture.createNewTmpDir();
-          fs.writeFileSync(
-            path.join(projectDir, "workflow_settings.yaml"),
-            dumpYaml(dataform.WorkflowSettings.create(testConfig))
-          );
-          fs.mkdirSync(path.join(projectDir, "definitions"));
-          fs.writeFileSync(path.join(projectDir, "definitions/e.sqlx"), `config {type: "view"}`);
-          fs.writeFileSync(path.join(projectDir, "definitions/file.sqlx"), "${resolve('e')}");
-
-          const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-          const suffix = testConfig.datasetSuffix ? `_${testConfig.datasetSuffix}` : "";
-          const prefix = testConfig.namePrefix ? `${testConfig.namePrefix}_` : "";
-          expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-          expect(result.compile.compiledGraph.operations[0].queries[0]).deep.equals(
-            `\`defaultDataset${suffix}.${prefix}e\``
-          );
-        });
-      });
-    });
-
-    suite("resolve with legacy dependencies", () => {
-      [
-        "assertion",
-        "incremental",
-        "operations",
-        "table",
-        "view",
-      ].forEach(actionType => {
-        test(`for action type: "${actionType}"`, () => {
-          const projectDir = tmpDirFixture.createNewTmpDir();
-          fs.writeFileSync(
-            path.join(projectDir, "workflow_settings.yaml"),
-            VALID_WORKFLOW_SETTINGS_YAML
-          );
-          fs.mkdirSync(path.join(projectDir, "definitions"));
-          fs.writeFileSync(path.join(projectDir, "definitions/source_a.sqlx"),
-            `
-config {
-type: "table",
-schema: "schema_a",
-name: "tbl",
-}
-
-SELECT 1`);
-          fs.writeFileSync(path.join(projectDir, "definitions/source_b.sqlx"),
-            `
-config {
-type: "table",
-schema: "schema_b",
-name: "tbl",
-}
-
-SELECT 1`);
-          fs.writeFileSync(path.join(projectDir, "definitions/file_with_dependencies.sqlx"),
-          `
-config {
-type: "${actionType}",
-dependencies: ["schema_a.tbl"]
-}
-
-SELECT 1`);        
-
-          const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-          expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-          const compiledAction = toVerifiableAction(result.compile.compiledGraph, actionType);
-          expect(asPlainObject(compiledAction)).deep.equals(
-            asPlainObject({
-              type: actionType,
-              target: {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "file_with_dependencies",
-              },
-              canonicalTarget: {
-                database:"defaultProject",
-                schema: "defaultDataset",
-                name: "file_with_dependencies"
-              },
-              dependencyTargets: [
-                {
-                  database: "defaultProject",
-                  schema: "schema_a",
-                  name: "tbl",
-                }
-              ],
-            })
-          );
-        });
-      });
-    });
-
-    test("fails when cannot resolve", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(path.join(projectDir, "definitions/file.sqlx"), "${resolve('e')}");
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(asPlainObject(result.compile.compiledGraph.operations[0].queries[0])).deep.equals(``);
-      expect(
-        asPlainObject(result.compile.compiledGraph.graphErrors.compilationErrors[0].message)
-      ).deep.equals(`Could not resolve "e"`);
-    });
-
-    test("failed ref attributes the compilation error to the source sqlx file", () => {
-      // Regression test: after the vm2 3.11.3 upgrade, V8 CallSite objects
-      // inside the sandbox lose their file paths, so utils.getCallerFile
-      // cannot recover the caller during a callback invocation. Errors from
-      // Session.resolve were therefore attributed to the sandbox entry
-      // ("index.js") instead of the action's source file. This test locks
-      // the fileName on the emitted CompilationError to the sqlx file.
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/mytable.sqlx"),
-        `config { type: "view" }\nSELECT 1 FROM \${ref("FOO")}`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      const errors = result.compile.compiledGraph.graphErrors.compilationErrors;
-      const resolveError = errors.find(e => e.message === `Could not resolve "FOO"`);
-      expect(resolveError, "expected a 'Could not resolve' error").to.not.equal(undefined);
-      expect(resolveError.fileName).equals("definitions/mytable.sqlx");
-    });
-
-    test("fails when ambiguous resolve", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/file.js"),
-        `
-publish("a", {"schema": "foo"})
-publish("a", {"schema": "bar"})
-publish("b", {"schema": "foo"}).dependencies("a")`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(
-        result.compile.compiledGraph.graphErrors.compilationErrors?.map(error => error.message)
-      ).deep.equals([
-        `Ambiguous Action name: {\"name\":\"a\",\"includeDependentAssertions\":false}. Did you mean one of: foo.a, bar.a.`
-      ]);
-    });
-
-    suite("context methods", () => {
-      [
-        WorkflowSettingsTemplates.bigqueryWithDefaultProjectAndDataset,
-        {
-          ...WorkflowSettingsTemplates.bigqueryWithDatasetSuffix,
-          defaultProject: "defaultProject"
-        },
-        { ...WorkflowSettingsTemplates.bigqueryWithNamePrefix, defaultProject: "defaultProject" }
-      ].forEach(testConfig => {
-        test(
-          `assertions target context functions with project suffix '${testConfig.projectSuffix}', ` +
-            `dataset suffix '${testConfig.datasetSuffix}', and name prefix '${testConfig.namePrefix}'`,
-          () => {
-            const projectDir = tmpDirFixture.createNewTmpDir();
-            fs.writeFileSync(
-              path.join(projectDir, "workflow_settings.yaml"),
-              dumpYaml(dataform.WorkflowSettings.create(testConfig))
-            );
-            fs.mkdirSync(path.join(projectDir, "definitions"));
-            fs.writeFileSync(
-              path.join(projectDir, "definitions/file.js"),
-              'assert("name", ctx => `${ctx.database()}.${ctx.schema()}.${ctx.name()}`)'
-            );
-
-            const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-            expect(
-              asPlainObject(result.compile.compiledGraph.graphErrors.compilationErrors)
-            ).deep.equals([]);
-            expect(asPlainObject(result.compile.compiledGraph.assertions[0].query)).deep.equals(
-              `defaultProject${testConfig.projectSuffix ? `_suffix` : ""}.` +
-                `defaultDataset${testConfig.datasetSuffix ? `_suffix` : ""}.` +
-                `${testConfig.namePrefix ? `prefix_` : ""}name`
-            );
-          }
-        );
-      });
-
-      test("assertions database function fails when database is undefined on the proto", () => {
-        const projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "workflow_settings.yaml"),
-          dumpYaml(dataform.WorkflowSettings.create(WorkflowSettingsTemplates.bigquery))
-        );
-        fs.mkdirSync(path.join(projectDir, "definitions"));
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/file.js"),
-          'assert("name", ctx => `${ctx.database()}.${ctx.schema()}.${ctx.name()}`)'
-        );
-
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        expect(
-          asPlainObject(result.compile.compiledGraph.graphErrors.compilationErrors?.[0]?.message)
-        ).deep.equals("Warehouse does not support multiple databases");
-      });
-    });
-
-    suite("filenames with multiple dots cause compilation errors", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(path.join(projectDir, "definitions/table1.extradot.sqlx"), "SELECT 1");
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- operation:
-    dataset: "dataset.extradot"
-    filename: table2.extradot.sql`
-      );
-      fs.writeFileSync(path.join(projectDir, "definitions/table2.extradot.sql"), "SELECT 2");
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(
-        result.compile.compiledGraph.graphErrors.compilationErrors
-          .map(({ message }) => message)
-          .sort()
-      ).deep.equals([
-        `Action target datasets cannot include '.'`,
-        `Action target datasets cannot include '.'`,
-        `Action target names cannot include '.'`,
-        `Action target names cannot include '.'`,
-        `Action target names cannot include '.'`
-      ]);
-    });
-
-    test("fails when non-unique target", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/file.js"),
-        `
-publish("name")
-publish("name")`
-      );
-
-      const result = runMainInVm(
-        coreExecutionRequestFromPath(
-          projectDir,
-          dataform.ProjectConfig.create({
-            defaultSchema: "otherDataset"
-          })
-        )
-      );
-
-      expect(
-        result.compile.compiledGraph.graphErrors.compilationErrors?.map(error => error.message)
-      ).deep.equals([
-        `Duplicate action name detected. Names within a schema must be unique across tables, declarations, assertions, and operations:\n\"{\"schema\":\"otherDataset\",\"name\":\"name\",\"database\":\"defaultProject\"}\"`,
-        `Duplicate canonical target detected. Canonical targets must be unique across tables, declarations, assertions, and operations:\n\"{\"schema\":\"otherDataset\",\"name\":\"name\",\"database\":\"defaultProject\"}\"`,
-        `Duplicate action name detected. Names within a schema must be unique across tables, declarations, assertions, and operations:\n\"{\"schema\":\"otherDataset\",\"name\":\"name\",\"database\":\"defaultProject\"}\"`,
-        `Duplicate canonical target detected. Canonical targets must be unique across tables, declarations, assertions, and operations:\n\"{\"schema\":\"otherDataset\",\"name\":\"name\",\"database\":\"defaultProject\"}\"`
-      ]);
-    });
-
-    test("fails when circular dependencies", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/file.js"),
-        `
-publish("a").dependencies("b")
-publish("b").dependencies("a")`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(
-        result.compile.compiledGraph.graphErrors.compilationErrors?.map(error => error.message)
-      ).deep.equals([
-        `Circular dependency detected in chain: [{\"database\":\"defaultProject\",\"name\":\"a\",\"schema\":\"defaultDataset\"} > {\"database\":\"defaultProject\",\"name\":\"b\",\"schema\":\"defaultDataset\"} > defaultProject.defaultDataset.a]`
-      ]);
-    });
-
-    test("fails when missing dependency", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(path.join(projectDir, "definitions/file.sql"), "unused");
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/file.js"),
-        `
-publish("a").dependencies("b")`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(
-        result.compile.compiledGraph.graphErrors.compilationErrors?.map(error => error.message)
-      ).deep.equals([
-        `Missing dependency detected: Action \"defaultProject.defaultDataset.a\" depends on \"{\"name\":\"b\",\"includeDependentAssertions\":false}\" which does not exist`
-      ]);
-    });
-
-    test("semi-colons at the end of SQL statements throws", () => {
-      // If this didn't happen, then the generated SQL could be incorrect
-      // because of being broken up by semi-colons.
-
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/file.js"),
-        `
-publish("a", "SELECT 1;\\n");
-publish("b", "SELECT 1;");`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(
-        result.compile.compiledGraph.graphErrors.compilationErrors?.map(error => error.message)
-      ).deep.equals([
-        "Semi-colons are not allowed at the end of SQL statements.",
-        "Semi-colons are not allowed at the end of SQL statements."
-      ]);
-    });
-  });
-
-  suite("sqlx special characters", () => {
-    test("extract blocks", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        dumpYaml(dataform.WorkflowSettings.create(WorkflowSettingsTemplates.bigquery))
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/file.sqlx"),
-        `
-config {
-  type: "table"
-}
-js {
-  var a = 1;
-}
-/*
-A multiline comment
-*/
-pre_operations {
-  SELECT 2;
-}
-post_operations {
-  SELECT 3;
-}
--- A single line comment.
-SELECT \${a}`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(result.compile.compiledGraph.tables[0].query).equals(`
-
-
-/*
-A multiline comment
-*/
-
-
--- A single line comment.
-SELECT 1`);
-      expect(result.compile.compiledGraph.tables[0].preOps[0]).equals(`
-  SELECT 2;
-`);
-      expect(result.compile.compiledGraph.tables[0].postOps[0]).equals(`
-  SELECT 3;
-`);
-    });
-
-    test("backticks appear to users as written", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        dumpYaml(dataform.WorkflowSettings.create(WorkflowSettingsTemplates.bigquery))
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      const fileContents = `select
-  "\`",
-  """\`"",
-from \`location\``;
-      fs.writeFileSync(path.join(projectDir, "definitions/file.sqlx"), fileContents);
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(result.compile.compiledGraph.operations[0].queries[0]).equals(fileContents);
-    });
-
-    test("backslashes appear to users as written", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        dumpYaml(dataform.WorkflowSettings.create(WorkflowSettingsTemplates.bigquery))
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      const sqlContents = `select
-  regexp_extract('01a_data_engine', '^(\\d{2}\\w)'),
-  regexp_extract('01a_data_engine', '^(\\\\d{2}\\\\w)'),
-  regexp_extract('\\\\', ''),
-  regexp_extract("", r"[0-9]\\"*"),
-  """\\ \\? \\\\"""`;
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/file.sqlx"),
-        `config { type: "table" }` + sqlContents + `pre_operations { ${sqlContents} }`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(result.compile.compiledGraph.tables[0].query.trim()).equals(sqlContents);
-      expect(result.compile.compiledGraph.tables[0].preOps[0].trim()).equals(sqlContents);
-    });
-
-    test("strings appear to users as written", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        dumpYaml(dataform.WorkflowSettings.create(WorkflowSettingsTemplates.bigquery))
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      const sqlContents = `select
-"""
-triple
-quotes
-""",
-"asd\\"123'def",
-'asd\\'123"def',
-
-select
-"""
-triple
-quotes
-""",
-"asd\\"123'def",
-'asd\\'123"def'`;
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/file.sqlx"),
-        `config { type: "table" }` + sqlContents + `post_operations { ${sqlContents} }`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(result.compile.compiledGraph.tables[0].query.trim()).equals(sqlContents);
-      expect(result.compile.compiledGraph.tables[0].postOps[0].trim()).equals(sqlContents);
-    });
-  });
-
-  suite("workflow settings", () => {
-    test(`valid workflow_settings.yaml is present`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.projectConfig)).deep.equals(
-        asPlainObject({
-          warehouse: "bigquery",
-          defaultDatabase: "defaultProject",
-          defaultSchema: "defaultDataset",
-          defaultLocation: "US"
-        })
-      );
-    });
-
-    // dataform.json for workflow settings is deprecated, but still currently supported.
-    test(`a valid dataform.json is present`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(path.join(projectDir, "dataform.json"), VALID_DATAFORM_JSON);
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.projectConfig)).deep.equals(
-        asPlainObject({
-          defaultDatabase: "defaultProject",
-          defaultLocation: "US",
-          defaultSchema: "defaultDataset"
-        })
-      );
-    });
-
-    test(`fails when no workflow settings file is present`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-
-      expect(() => runMainInVm(coreExecutionRequestFromPath(projectDir))).to.throw(
-        "Failed to resolve workflow_settings.yaml"
-      );
-    });
-
-    test(`fails when both workflow settings and dataform.json files are present`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(path.join(projectDir, "dataform.json"), VALID_DATAFORM_JSON);
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-
-      expect(() => runMainInVm(coreExecutionRequestFromPath(projectDir))).to.throw(
-        "dataform.json has been deprecated and cannot be defined alongside workflow_settings.yaml"
-      );
-    });
-
-    test(`fails when workflow_settings.yaml cannot be represented in JSON format`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(path.join(projectDir, "workflow_settings.yaml"), "&*19132sdS:asd:");
-
-      expect(() => runMainInVm(coreExecutionRequestFromPath(projectDir))).to.throw(
-        "workflow_settings.yaml is invalid"
-      );
-    });
-
-    test(`fails when workflow settings fails to be parsed`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        `
-someKey: and an extra: colon
-`
-      );
-
-      expect(() => runMainInVm(coreExecutionRequestFromPath(projectDir))).to.throw(
-        "workflow_settings.yaml is not a valid YAML file: YAMLException: bad indentation"
-      );
-    });
-
-    test(`fails when dataform.json is an invalid json file`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(path.join(projectDir, "dataform.json"), '{keyWithNoQuotes: "validValue"}');
-
-      expect(() => runMainInVm(coreExecutionRequestFromPath(projectDir))).to.throw(
-        "Expected property name or '}' in JSON at position 1 (line 1 column 2)"
-      );
-    });
-
-    test(`fails when a valid workflow_settings.yaml contains unknown fields`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        "notAProjectConfigField: value"
-      );
-
-      expect(() => runMainInVm(coreExecutionRequestFromPath(projectDir))).to.throw(
-        `Workflow settings error: Unexpected property "notAProjectConfigField", or property value type of "string" is incorrect. See https://dataform-co.github.io/dataform/docs/configs-reference#dataform-WorkflowSettings for allowed properties.`
-      );
-    });
-
-    test(`fails when a valid workflow_settings.yaml base level is an array`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(path.join(projectDir, "workflow_settings.yaml"), "- someArrayEntry");
-
-      expect(() => runMainInVm(coreExecutionRequestFromPath(projectDir))).to.throw(
-        "Expected a top-level object, but found an array"
-      );
-    });
-
-    test(`fails when a valid dataform.json contains unknown fields`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "dataform.json"),
-        `{"notAProjectConfigField": "value"}`
-      );
-
-      expect(() => runMainInVm(coreExecutionRequestFromPath(projectDir))).to.throw(
-        `Dataform json error: Unexpected property "notAProjectConfigField", or property value type of "string" is incorrect.`
-      );
-    });
-
-    test(`does not fail when defaultLocation is not present in workflow_settings.yaml`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        `
-dataformCoreVersion: ${version}
-defaultProject: project`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.projectConfig)).deep.equals(
-        asPlainObject({
-          warehouse: "bigquery",
-          defaultDatabase: "project"
-        })
-      );
-    });
-
-    test(`workflow settings and project config overrides are merged and applied within SQLX files`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        `
-defaultProject: defaultProject
-defaultLocation: locationInWorkflowSettings
-vars:
-  selectVar: selectVal
-`
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/file.sqlx"),
-        `
-config {
-  type: "table",
-  database: dataform.projectConfig.vars.projectVar,
-}
-select 1 AS \${dataform.projectConfig.vars.selectVar}`
-      );
-      const coreExecutionRequest = dataform.CoreExecutionRequest.create({
-        compile: {
-          compileConfig: {
-            projectDir,
-            filePaths: ["definitions/file.sqlx"],
-            projectConfigOverride: {
-              defaultLocation: "locationInOverride",
-              vars: {
-                projectVar: "projectVal"
-              }
-            }
-          }
-        }
-      });
-
-      const result = runMainInVm(coreExecutionRequest);
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph)).deep.equals(
-        asPlainObject({
-          dataformCoreVersion: version,
-          graphErrors: {},
-          jitData: {},
-          projectConfig: {
-            defaultDatabase: "defaultProject",
-            defaultLocation: "locationInOverride",
-            vars: {
-              projectVar: "projectVal",
-              selectVar: "selectVal"
-            },
-            warehouse: "bigquery"
-          },
-          tables: [
-            {
-              canonicalTarget: {
-                database: "projectVal",
-                name: "file"
-              },
-              disabled: false,
-              enumType: "TABLE",
-              fileName: "definitions/file.sqlx",
-              hermeticity: "NON_HERMETIC",
-              query: "\n\nselect 1 AS selectVal",
-              target: {
-                database: "projectVal",
-                name: "file"
-              },
-              type: "table"
-            }
-          ],
-          targets: [
-            {
-              database: "projectVal",
-              name: "file"
-            }
-          ]
-        })
-      );
-    });
-
-    suite("dataform core version", () => {
-      test(`main fails when the workflow settings version is not the installed current version`, () => {
-        const projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "workflow_settings.yaml"),
-          `
-dataformCoreVersion: 1.0.0
-defaultProject: dataform`
-        );
-
-        expect(() => runMainInVm(coreExecutionRequestFromPath(projectDir))).to.throw(
-          `Version mismatch: workflow settings specifies version 1.0.0, but ${version} was found`
-        );
-      });
-
-      test(`main succeeds when workflow settings contains the matching version`, () => {
-        const projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "workflow_settings.yaml"),
-          `
-dataformCoreVersion: ${version}
-defaultProject: project
-defaultLocation: US`
-        );
-
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-        expect(asPlainObject(result.compile.compiledGraph.projectConfig)).deep.equals(
-          asPlainObject({
-            warehouse: "bigquery",
-            defaultDatabase: "project",
-            defaultLocation: "US"
-          })
-        );
-      });
-    });
-
-    suite("variables", () => {
-      test(`variables in workflow_settings.yaml must be strings`, () => {
-        const projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "workflow_settings.yaml"),
-          `
-vars:
-  intValue: 1
-  strValue: "str"`
-        );
-
-        expect(() => runMainInVm(coreExecutionRequestFromPath(projectDir))).to.throw(
-          "Custom variables defined in workflow settings can only be strings."
-        );
-      });
-
-      test(`variables in dataform.json must be strings`, () => {
-        const projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "dataform.json"),
-          `{"vars": { "intVar": 1, "strVar": "str" } }`
-        );
-
-        expect(() => runMainInVm(coreExecutionRequestFromPath(projectDir))).to.throw(
-          "Custom variables defined in workflow settings can only be strings."
-        );
-      });
-
-      test(`variables can be referenced in SQLX`, () => {
-        const projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "workflow_settings.yaml"),
-          `
-defaultLocation: "us"
-vars:
-  descriptionVar: descriptionValue
-  columnVar: columnValue`
-        );
-        fs.mkdirSync(path.join(projectDir, "definitions"));
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/file.sqlx"),
-          // TODO(https://github.com/dataform-co/dataform/issues/1295): add a test and fix
-          // functionality for assertions overriding database.
-          `
-config {
-  type: "table",
-  database: dataform.projectConfig.vars.databaseVar,
-  schema: "tableSchema",
-  description: dataform.projectConfig.vars.descriptionVar,
-  assertions: {
-    nonNull: [dataform.projectConfig.vars.columnVar],
-  }
-}
-select 1 AS \${dataform.projectConfig.vars.columnVar}`
-        );
-        const coreExecutionRequest = dataform.CoreExecutionRequest.create({
-          compile: {
-            compileConfig: {
-              projectDir,
-              filePaths: ["definitions/file.sqlx"],
-              projectConfigOverride: {
-                vars: {
-                  databaseVar: "databaseVal"
-                }
-              }
-            }
-          }
-        });
-
-        const result = runMainInVm(coreExecutionRequest);
-
-        expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-        expect(asPlainObject(result.compile.compiledGraph)).deep.equals(
-          asPlainObject({
-            assertions: [
-              {
-                canonicalTarget: {
-                  name: "tableSchema_file_assertions_rowConditions"
-                },
-                dependencyTargets: [
-                  {
-                    database: "databaseVal",
-                    name: "file",
-                    schema: "tableSchema"
-                  }
-                ],
-                fileName: "definitions/file.sqlx",
-                parentAction: {
-                  database: "databaseVal",
-                  name: "file",
-                  schema: "tableSchema"
-                },
-                query:
-                  "\nSELECT\n  'columnValue IS NOT NULL' AS failing_row_condition,\n  *\nFROM `databaseVal.tableSchema.file`\nWHERE NOT (columnValue IS NOT NULL)\n",
-                target: {
-                  name: "tableSchema_file_assertions_rowConditions"
-                }
-              }
-            ],
-            dataformCoreVersion: version,
-            graphErrors: {},
-            jitData: {},
-            projectConfig: {
-              defaultLocation: "us",
-              vars: {
-                databaseVar: "databaseVal",
-                descriptionVar: "descriptionValue",
-                columnVar: "columnValue"
-              },
-              warehouse: "bigquery"
-            },
-            tables: [
-              {
-                actionDescriptor: {
-                  description: "descriptionValue"
-                },
-                canonicalTarget: {
-                  database: "databaseVal",
-                  name: "file",
-                  schema: "tableSchema"
-                },
-                disabled: false,
-                enumType: "TABLE",
-                fileName: "definitions/file.sqlx",
-                query: "\n\nselect 1 AS columnValue",
-                target: {
-                  database: "databaseVal",
-                  name: "file",
-                  schema: "tableSchema"
-                },
-                type: "table",
-                hermeticity: "NON_HERMETIC"
-              }
-            ],
-            targets: [
-              {
-                name: "tableSchema_file_assertions_rowConditions"
-              },
-              {
-                database: "databaseVal",
-                name: "file",
-                schema: "tableSchema"
-              }
-            ]
-          })
-        );
-      });
-    });
-  });
-
-  suite("action configs", () => {
-    test(`fails when file is not found`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- operation:
-    filename: doesnotexist.sql`
-      );
-
-      expect(() => runMainInVm(coreExecutionRequestFromPath(projectDir))).to.throw(
-        "Cannot find module 'definitions/doesnotexist.sql'"
-      );
-    });
-
-    test(`fails when properties belonging to other action config types are populated for an action config`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- table:
-    filename: action.sql
-    materialized: true`
-      );
-
-      expect(() => runMainInVm(coreExecutionRequestFromPath(projectDir))).to.throw(
-        `Unexpected property "materialized", or property value type of "boolean" is incorrect. See https://dataform-co.github.io/dataform/docs/configs-reference#dataform-ActionConfigs for allowed properties.`
-      );
-    });
-
-    test(`fails when empty objects are given`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:`
-      );
-
-      expect(() => runMainInVm(coreExecutionRequestFromPath(projectDir))).to.throw(
-        `Unexpected empty value for "actions". See https://dataform-co.github.io/dataform/docs/configs-reference#dataform-ActionConfigs for allowed properties.`
-      );
-    });
-
-    suite(`.sqlx filenames in actions.yaml are rejected with a clear error`, () => {
-      [
-        {
-          actionType: "table",
-          yamlBody: `
-actions:
-- table:
-    filename: table.sqlx`
-        },
-        {
-          actionType: "view",
-          yamlBody: `
-actions:
-- view:
-    filename: view.sqlx`
-        },
-        {
-          actionType: "incrementalTable",
-          yamlBody: `
-actions:
-- incrementalTable:
-    filename: incremental.sqlx`
-        },
-        {
-          actionType: "assertion",
-          yamlBody: `
-actions:
-- assertion:
-    filename: assertion.sqlx`
-        },
-        {
-          actionType: "operation",
-          yamlBody: `
-actions:
-- operation:
-    filename: operation.sqlx`
-        },
-        {
-          actionType: "declaration",
-          yamlBody: `
-actions:
-- declaration:
-    name: declaration
-    filename: declaration.sqlx`
-        },
-        {
-          actionType: "notebook",
-          yamlBody: `
-actions:
-- notebook:
-    filename: notebook.sqlx`
-        }
-      ].forEach(({ actionType, yamlBody }) => {
-        test(`for action type "${actionType}"`, () => {
-          const projectDir = tmpDirFixture.createNewTmpDir();
-          fs.writeFileSync(
-            path.join(projectDir, "workflow_settings.yaml"),
-            VALID_WORKFLOW_SETTINGS_YAML
-          );
-          fs.mkdirSync(path.join(projectDir, "definitions"));
-          fs.writeFileSync(path.join(projectDir, "definitions/actions.yaml"), yamlBody);
-
-          const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-          const errorMessages = result.compile.compiledGraph.graphErrors.compilationErrors.map(
-            ({ message }) => message
-          );
-          expect(errorMessages).to.have.lengthOf(1);
-          expect(errorMessages[0]).to.include(`Action config "${actionType}" has filename`);
-          expect(errorMessages[0]).to.include(
-            ".sqlx files cannot be referenced from actions.yaml"
-          );
-        });
-      });
-
-      test(`is case-insensitive`, () => {
-        const projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "workflow_settings.yaml"),
-          VALID_WORKFLOW_SETTINGS_YAML
-        );
-        fs.mkdirSync(path.join(projectDir, "definitions"));
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/actions.yaml"),
-          `
-actions:
-- table:
-    filename: table.SQLX`
-        );
-
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        expect(
-          result.compile.compiledGraph.graphErrors.compilationErrors.map(({ message }) => message)
-        ).to.have.lengthOf(1);
-      });
-
-      test(`for data preparations referencing a .dp.sqlx file`, () => {
-        // .dp.sqlx data preparations are compiled directly from the definitions/
-        // directory, so referencing one from actions.yaml is the same mistake as
-        // for any other action type and must produce the same clear error rather
-        // than silently loading a malformed, duplicated action.
-        const projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "workflow_settings.yaml"),
-          VALID_WORKFLOW_SETTINGS_YAML
-        );
-        fs.mkdirSync(path.join(projectDir, "definitions"));
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/actions.yaml"),
-          `
-actions:
-- dataPreparation:
-    filename: prep.dp.sqlx`
-        );
-
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        const errorMessages = result.compile.compiledGraph.graphErrors.compilationErrors.map(
-          ({ message }) => message
-        );
-        expect(errorMessages).to.have.lengthOf(1);
-        expect(errorMessages[0]).to.include(`Action config "dataPreparation" has filename`);
-        expect(errorMessages[0]).to.include(".sqlx files cannot be referenced from actions.yaml");
-      });
-    });
-
-    test(`filenames with non-UTF8 characters are valid`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- operation:
-    filename: utf8characters:私🙂 and some spaces.sql`
-      );
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/utf8characters:私🙂 and some spaces.sql"),
-        "SELECT 1"
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.operations)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "utf8characters:私🙂 and some spaces"
-            },
-            canonicalTarget: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "utf8characters:私🙂 and some spaces"
-            },
-            fileName: "definitions/utf8characters:私🙂 and some spaces.sql",
-            queries: ["SELECT 1"],
-            hermeticity: "NON_HERMETIC"
-          }
-        ])
-      );
-    });
-
-    test(`dependency targets are loaded`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- declaration:
-    name: declaration
-- table:
-    filename: table.sql
-    dependencyTargets:
-    - name: declaration
-      dataset: defaultDataset
-      project: defaultProject
-- incrementalTable:
-    filename: incrementalTable.sql
-    dependencyTargets:
-    - name: table
-      dataset: defaultDataset
-      project: defaultProject
-- view:
-    filename: view.sql
-    dependencyTargets:
-    - name: incrementalTable
-      dataset: defaultDataset
-      project: defaultProject
-- operation:
-    filename: operation.sql
-    dependencyTargets:
-    - name: view
-      dataset: defaultDataset
-      project: defaultProject
-- notebook:
-    filename: notebook.ipynb
-    dependencyTargets:
-    - name: view
-      dataset: defaultDataset
-      project: defaultProject`
-      );
-      ["table.sql", "incrementalTable.sql", "view.sql", "operation.sql"].forEach(filename => {
-        fs.writeFileSync(path.join(projectDir, `definitions/${filename}`), "SELECT 1");
-      });
-      fs.writeFileSync(
-        path.join(projectDir, `definitions/notebook.ipynb`),
-        EMPTY_NOTEBOOK_CONTENTS
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-    });
-
-    test(`dependency targets of actions with different types are loaded`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      // The dependency target for depending on a notebook currently hacks around the limitations of
-      // the target proto, until proper target support for notebooks is added.
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- notebook:
-    name: notebook1
-    location: location
-    project: project
-    filename: notebook.ipynb
-- operation:
-    name: operation1
-    dataset: dataset
-    project: project
-    dependencyTargets:
-    - name: notebook1
-      dataset: location
-      project: project
-    filename: operation.sql`
-      );
-      fs.writeFileSync(path.join(projectDir, "definitions/operation.sql"), "SELECT 1");
-      fs.writeFileSync(
-        path.join(projectDir, `definitions/notebook.ipynb`),
-        EMPTY_NOTEBOOK_CONTENTS
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-    });
-  });
-
-  suite("javascript API", () => {
-    suite("publish", () => {
-      ["table", "view", "incremental"].forEach(tableType => {
-        [
-          WorkflowSettingsTemplates.bigqueryWithDefaultProjectAndDataset,
-          {
-            ...WorkflowSettingsTemplates.bigqueryWithDatasetSuffix,
-            defaultProject: "defaultProject"
-          },
-          { ...WorkflowSettingsTemplates.bigqueryWithNamePrefix, defaultProject: "defaultProject" }
-        ].forEach(projectConfig => {
-          test(
-            `publish for table type ${tableType}, with project suffix ` +
-              `'${projectConfig.projectSuffix}', dataset suffix ` +
-              `'${projectConfig.datasetSuffix}', and name prefix '${projectConfig.namePrefix}'`,
-            () => {
-              const projectDir = tmpDirFixture.createNewTmpDir();
-              fs.writeFileSync(
-                path.join(projectDir, "workflow_settings.yaml"),
-                dumpYaml(dataform.WorkflowSettings.create(projectConfig))
-              );
-              fs.mkdirSync(path.join(projectDir, "definitions"));
-              fs.writeFileSync(
-                path.join(projectDir, "definitions/publish.js"),
-                `
-publish("name", {
-  type: "${tableType}",
-}).query(_ => "SELECT 1")
-  .preOps(_ => ["pre_op"])
-  .postOps(_ => ["post_op"])
-  .database("otherProject")
-  .schema("otherDataset")`
-              );
-
-              const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-              expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-              expect(asPlainObject(result.compile.compiledGraph.tables)).deep.equals(
-                asPlainObject([
-                  {
-                    type: tableType,
-                    hermeticity: "NON_HERMETIC",
-                    target: {
-                      database: projectConfig.projectSuffix
-                        ? `otherProject_${projectConfig.projectSuffix}`
-                        : "otherProject",
-                      schema: projectConfig.datasetSuffix
-                        ? `otherDataset_${projectConfig.datasetSuffix}`
-                        : "otherDataset",
-                      name: projectConfig.namePrefix ? `${projectConfig.namePrefix}_name` : "name"
-                    },
-                    canonicalTarget: {
-                      database: projectConfig.defaultProject,
-                      schema: projectConfig.defaultDataset,
-                      name: "name"
-                    },
-                    disabled: false,
-                    enumType: tableType.toUpperCase(),
-                    fileName: "definitions/publish.js",
-                    query: "SELECT 1",
-                    postOps: ["post_op"],
-                    preOps: ["pre_op"],
-                    ...(tableType === "incremental"
-                      ? {
-                          incrementalPostOps: ["post_op"],
-                          incrementalPreOps: ["pre_op"],
-                          incrementalQuery: "SELECT 1",
-                          incrementalStrategy: "INCREMENTAL_STRATEGY_UNSPECIFIED",
-                          protected: false,
-                          onSchemaChange: "IGNORE"
-                        }
-                      : {})
-                  }
-                ])
-              );
-            }
-          );
-        });
-
-        test("ref resolved correctly", () => {
-          const projectDir = tmpDirFixture.createNewTmpDir();
-          fs.writeFileSync(
-            path.join(projectDir, "workflow_settings.yaml"),
-            VALID_WORKFLOW_SETTINGS_YAML
-          );
-          fs.mkdirSync(path.join(projectDir, "definitions"));
-          fs.writeFileSync(
-            path.join(projectDir, "definitions/operation.sqlx"),
-            `
-config {
-  hasOutput: true
-}
-SELECT 1`
-          );
-          fs.writeFileSync(
-            path.join(projectDir, "definitions/publish.js"),
-            `
-publish("name", {
-  type: "${tableType}",
-}).query(ctx => \`SELECT * FROM \${ctx.ref('operation')}\`)`
-          );
-
-          const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-          expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-          expect(asPlainObject(result.compile.compiledGraph.tables)).deep.equals([
-            {
-              canonicalTarget: {
-                database: "defaultProject",
-                name: "name",
-                schema: "defaultDataset"
-              },
-              dependencyTargets: [
-                {
-                  database: "defaultProject",
-                  name: "operation",
-                  schema: "defaultDataset"
-                }
-              ],
-              disabled: false,
-              enumType: tableType.toUpperCase(),
-              fileName: "definitions/publish.js",
-              hermeticity: "NON_HERMETIC",
-              query: "SELECT * FROM `defaultProject.defaultDataset.operation`",
-              target: {
-                database: "defaultProject",
-                name: "name",
-                schema: "defaultDataset"
-              },
-              type: tableType,
-              ...(tableType === "incremental"
-                ? {
-                    incrementalQuery: "SELECT * FROM `defaultProject.defaultDataset.operation`",
-                    incrementalStrategy: "INCREMENTAL_STRATEGY_UNSPECIFIED",
-                    protected: false,
-                    onSchemaChange: "IGNORE"
-                  }
-                : {})
-            }
-          ]);
-        });
-
-        if (tableType === "incremental") {
-          test("protected resolved correctly", () => {
-            const projectDir = tmpDirFixture.createNewTmpDir();
-            fs.writeFileSync(
-              path.join(projectDir, "workflow_settings.yaml"),
-              VALID_WORKFLOW_SETTINGS_YAML
-            );
-            fs.mkdirSync(path.join(projectDir, "definitions"));
-            fs.writeFileSync(
-              path.join(projectDir, "definitions/publish.js"),
-              `
-publish("name", {
-  type: "incremental",
-}).query(_ => "SELECT 1")
-  .protected()`
-            );
-  
-            const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-  
-            expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-            expect(asPlainObject(result.compile.compiledGraph.tables)).deep.equals([
-              {
-                canonicalTarget: {
-                  database: "defaultProject",
-                  name: "name",
-                  schema: "defaultDataset"
-                },
-                disabled: false,
-                enumType: tableType.toUpperCase(),
-                fileName: "definitions/publish.js",
-                hermeticity: "NON_HERMETIC",
-                query: "SELECT 1",
-                target: {
-                  database: "defaultProject",
-                  name: "name",
-                  schema: "defaultDataset"
-                },
-                type: tableType,
-                incrementalQuery: "SELECT 1",
-                protected: true,
-                onSchemaChange: "IGNORE",
-                incrementalStrategy: "INCREMENTAL_STRATEGY_UNSPECIFIED",
-              }
-            ]);
-          }); 
-        }
-      });
-    });
-
-    suite("operate", () => {
-      [
-        WorkflowSettingsTemplates.bigqueryWithDefaultProjectAndDataset,
-        {
-          ...WorkflowSettingsTemplates.bigqueryWithDatasetSuffix,
-          defaultProject: "defaultProject"
-        },
-        { ...WorkflowSettingsTemplates.bigqueryWithNamePrefix, defaultProject: "defaultProject" }
-      ].forEach(projectConfig => {
-        test(
-          `operate with project suffix ` +
-            `'${projectConfig.projectSuffix}', dataset suffix ` +
-            `'${projectConfig.datasetSuffix}', and name prefix '${projectConfig.namePrefix}'`,
-          () => {
-            const projectDir = tmpDirFixture.createNewTmpDir();
-            fs.writeFileSync(
-              path.join(projectDir, "workflow_settings.yaml"),
-              dumpYaml(dataform.WorkflowSettings.create(projectConfig))
-            );
-            fs.mkdirSync(path.join(projectDir, "definitions"));
-            fs.writeFileSync(
-              path.join(projectDir, "definitions/operate.js"),
-              `
-operate("name", {
-  type: "operations",
-}).queries(_ => ["SELECT 1", "SELECT 2"])
-  .database("otherProject")
-  .schema("otherDataset")`
-            );
-
-            const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-            expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-            expect(asPlainObject(result.compile.compiledGraph.operations)).deep.equals(
-              asPlainObject([
-                {
-                  target: {
-                    database: projectConfig.projectSuffix
-                      ? `otherProject_${projectConfig.projectSuffix}`
-                      : "otherProject",
-                    schema: projectConfig.datasetSuffix
-                      ? `otherDataset_${projectConfig.datasetSuffix}`
-                      : "otherDataset",
-                    name: projectConfig.namePrefix ? `${projectConfig.namePrefix}_name` : "name"
-                  },
-                  canonicalTarget: {
-                    database: projectConfig.defaultProject,
-                    schema: projectConfig.defaultDataset,
-                    name: "name"
-                  },
-                  fileName: "definitions/operate.js",
-                  hermeticity: "NON_HERMETIC",
-                  queries: ["SELECT 1", "SELECT 2"]
-                }
-              ])
-            );
-          }
-        );
-      });
-
-      test("ref resolved correctly", () => {
-        const projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "workflow_settings.yaml"),
-          VALID_WORKFLOW_SETTINGS_YAML
-        );
-        fs.mkdirSync(path.join(projectDir, "definitions"));
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/table.sqlx"),
-          `config {type: "table"} SELECT 1`
-        );
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/operate.js"),
-          `
-operate("name", {
-  type: "operations",
-}).queries(ctx => [\`SELECT * FROM \${ctx.ref('table')}\`])`
-        );
-
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-        expect(asPlainObject(result.compile.compiledGraph.operations)).deep.equals([
-          {
-            canonicalTarget: {
-              database: "defaultProject",
-              name: "name",
-              schema: "defaultDataset"
-            },
-            dependencyTargets: [
-              {
-                database: "defaultProject",
-                name: "table",
-                schema: "defaultDataset"
-              }
-            ],
-            fileName: "definitions/operate.js",
-            hermeticity: "NON_HERMETIC",
-            queries: ["SELECT * FROM `defaultProject.defaultDataset.table`"],
-            target: {
-              database: "defaultProject",
-              name: "name",
-              schema: "defaultDataset"
-            }
-          }
-        ]);
-      });
-    });
-
-    suite("assert", () => {
-      [
-        WorkflowSettingsTemplates.bigqueryWithDefaultProjectAndDataset,
-        {
-          ...WorkflowSettingsTemplates.bigqueryWithDatasetSuffix,
-          defaultProject: "defaultProject"
-        },
-        { ...WorkflowSettingsTemplates.bigqueryWithNamePrefix, defaultProject: "defaultProject" }
-      ].forEach(projectConfig => {
-        test(
-          `assert with project suffix ` +
-            `'${projectConfig.projectSuffix}', dataset suffix ` +
-            `'${projectConfig.datasetSuffix}', and name prefix '${projectConfig.namePrefix}'`,
-          () => {
-            const projectDir = tmpDirFixture.createNewTmpDir();
-            fs.writeFileSync(
-              path.join(projectDir, "workflow_settings.yaml"),
-              dumpYaml(dataform.WorkflowSettings.create(projectConfig))
-            );
-            fs.mkdirSync(path.join(projectDir, "definitions"));
-            fs.writeFileSync(
-              path.join(projectDir, "definitions/assert.js"),
-              `
-assert("name", {
-  type: "operations",
-}).query(_ => "SELECT 1")
-  .database("otherProject")
-  .schema("otherDataset")`
-            );
-
-            const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-            expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-            expect(asPlainObject(result.compile.compiledGraph.assertions)).deep.equals(
-              asPlainObject([
-                {
-                  target: {
-                    database: projectConfig.projectSuffix
-                      ? `otherProject_${projectConfig.projectSuffix}`
-                      : "otherProject",
-                    schema: projectConfig.datasetSuffix
-                      ? `otherDataset_${projectConfig.datasetSuffix}`
-                      : "otherDataset",
-                    name: projectConfig.namePrefix ? `${projectConfig.namePrefix}_name` : "name"
-                  },
-                  canonicalTarget: {
-                    database: projectConfig.defaultProject,
-                    schema: projectConfig.defaultDataset,
-                    name: "name"
-                  },
-                  fileName: "definitions/assert.js",
-                  query: "SELECT 1"
-                }
-              ])
-            );
-          }
-        );
-      });
-
-      test("ref resolved correctly", () => {
-        const projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "workflow_settings.yaml"),
-          VALID_WORKFLOW_SETTINGS_YAML
-        );
-        fs.mkdirSync(path.join(projectDir, "definitions"));
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/table.sqlx"),
-          `config {type: "table"} SELECT 1`
-        );
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/assert.js"),
-          `
-assert("name", {
-  type: "assert",
-}).query(ctx => \`SELECT * FROM \${ctx.ref('table')}\`)`
-        );
-
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-        expect(asPlainObject(result.compile.compiledGraph.assertions)).deep.equals([
-          {
-            canonicalTarget: {
-              database: "defaultProject",
-              name: "name",
-              schema: "defaultDataset"
-            },
-            dependencyTargets: [
-              {
-                database: "defaultProject",
-                name: "table",
-                schema: "defaultDataset"
-              }
-            ],
-            fileName: "definitions/assert.js",
-            query: "SELECT * FROM `defaultProject.defaultDataset.table`",
-            target: {
-              database: "defaultProject",
-              name: "name",
-              schema: "defaultDataset"
-            }
-          }
-        ]);
-      });
-
-      test("jitCode correctly populates the jitCode field", () => {
-        const projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "workflow_settings.yaml"),
-          VALID_WORKFLOW_SETTINGS_YAML
-        );
-        fs.mkdirSync(path.join(projectDir, "definitions"));
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/assert.js"),
-          `
-assert("name").jitCode(ctx => "jit");`
-        );
-
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-        expect(asPlainObject(result.compile.compiledGraph.assertions)).deep.equals([
-          {
-            actionDescriptor: {
-              compilationMode: "ACTION_COMPILATION_MODE_JIT"
-            },
-            canonicalTarget: {
-              database: "defaultProject",
-              name: "name",
-              schema: "defaultDataset"
-            },
-            fileName: "definitions/assert.js",
-            jitCode: 'ctx => "jit"',
-            target: {
-              database: "defaultProject",
-              name: "name",
-              schema: "defaultDataset"
-            }
-          }
-        ]);
-      });
-
-      test("fails when both jitCode and query are set on an assertion", () => {
-        const projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "workflow_settings.yaml"),
-          VALID_WORKFLOW_SETTINGS_YAML
-        );
-        fs.mkdirSync(path.join(projectDir, "definitions"));
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/assert.js"),
-          `
-assert("name").query("SELECT 1").jitCode(ctx => "jit");`
-        );
-
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        expect(
-          result.compile.compiledGraph.graphErrors.compilationErrors?.map(error => error.message)
-        ).deep.equals([
-          "Assertion may set either .jitCode() or .query(), but not both."
-        ]);
-      });
-
-      test("assert API returns disabled assertions when disableAssertions is true", () => {
-        const projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "workflow_settings.yaml"),
-          VALID_WORKFLOW_SETTINGS_YAML
-        );
-        fs.mkdirSync(path.join(projectDir, "definitions"));
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/table.sqlx"),
-          `config {type: "table"} SELECT 1`
-        );
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/assert.js"),
-          `
-assert("name", {
-  type: "assert",
-}).query(ctx => \`SELECT * FROM \${ctx.ref('table')}\`)`
-        );
-
-        const coreRequest = coreExecutionRequestFromPath(
-          projectDir,
-          dataform.ProjectConfig.create({
-            disableAssertions: true
-          })
-        );
-        const result = runMainInVm(coreRequest);
-
-        expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-        expect(asPlainObject(result.compile.compiledGraph.assertions)).deep.equals(
-          asPlainObject([
-            {
-              canonicalTarget: {
-                database: "defaultProject",
-                name: "name",
-                schema: "defaultDataset"
-              },
-              dependencyTargets: [
-                {
-                  database: "defaultProject",
-                  name: "table",
-                  schema: "defaultDataset"
-                }
-              ],
-              disabled: true,
-              fileName: "definitions/assert.js",
-              query: "SELECT * FROM `defaultProject.defaultDataset.table`",
-              target: {
-                database: "defaultProject",
-                name: "name",
-                schema: "defaultDataset"
-              }
-            }
-          ])
-        );
-        expect(result.compile.compiledGraph.tables.length).equals(1);
-      });
-    });
-
-    suite("jitData", () => {
-      test("jitData is added to the compiled graph", () => {
-        const projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "workflow_settings.yaml"),
-          VALID_WORKFLOW_SETTINGS_YAML
-        );
-        fs.mkdirSync(path.join(projectDir, "definitions"));
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/jit.js"),
-          `
-dataform.jitData("key", {
-  "number": 123,
-  "string": "value",
-  "boolean": true,
-  "struct": {
-    "nestedKey": "nestedValue"
-  },
-  "list": [
-    "a",
-    "b",
-    "c"
-  ],
-  "null": null,
-  "undef": undefined,
-});`
-        );
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-        expect(result.compile.compiledGraph.jitData).to.deep.equal(
-          google.protobuf.Struct.create({
-            fields: {
-              key: google.protobuf.Value.create({
-                structValue: google.protobuf.Struct.create({
-                  fields: {
-                    number: google.protobuf.Value.create({ numberValue: 123 }),
-                    string: google.protobuf.Value.create({ stringValue: "value" }),
-                    boolean: google.protobuf.Value.create({ boolValue: true }),
-                    struct: google.protobuf.Value.create({
-                      structValue: google.protobuf.Struct.create({
-                        fields: {
-                          nestedKey: google.protobuf.Value.create({ stringValue: "nestedValue" })
-                        }
-                      })
-                    }),
-                    list: google.protobuf.Value.create({
-                      listValue: google.protobuf.ListValue.create({
-                        values: [
-                          google.protobuf.Value.create({ stringValue: "a" }),
-                          google.protobuf.Value.create({ stringValue: "b" }),
-                          google.protobuf.Value.create({ stringValue: "c" })
-                        ]
-                      })
-                    }),
-                    null: google.protobuf.Value.create({
-                      nullValue: google.protobuf.NullValue.NULL_VALUE
-                    }),
-                    undef: google.protobuf.Value.create({
-                      nullValue: google.protobuf.NullValue.NULL_VALUE
-                    }),
-                  }
-                })
-              })
-            }
-          })
-        );
-      });
-
-      test("jitData with duplicate key throws error", () => {
-        const projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "workflow_settings.yaml"),
-          VALID_WORKFLOW_SETTINGS_YAML
-        );
-        fs.mkdirSync(path.join(projectDir, "definitions"));
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/jit.js"),
-          `
-dataform.jitData("key", 1);
-dataform.jitData("key", 2);
-`
-        );
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        expect(
-          result.compile.compiledGraph.graphErrors.compilationErrors.map(e => e.message)
-        ).to.deep.equal(["JiT context data with key key already exists."]);
-      });
-
-      test("jitData with unsupported type throws error", () => {
-        const projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "workflow_settings.yaml"),
-          VALID_WORKFLOW_SETTINGS_YAML
-        );
-        fs.mkdirSync(path.join(projectDir, "definitions"));
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/jit.js"),
-          `
-dataform.jitData("key", {test: () => {}});
-`
-        );
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        expect(
-          result.compile.compiledGraph.graphErrors.compilationErrors.map(e => e.message)
-        ).to.deep.equal(["Unsupported value: () => {}"]);
-      });
-    });
-    suite("markdown in description", () => {
-      test("markdown contents added to description in sqlx", () => {
-          const projectDir = tmpDirFixture.createNewTmpDir();
-          fs.writeFileSync(
-            path.join(projectDir, "workflow_settings.yaml"),
-            VALID_WORKFLOW_SETTINGS_YAML
-          );
-          fs.mkdirSync(path.join(projectDir, "definitions"));
-          fs.writeFileSync(
-            path.join(projectDir, "definitions/descriptions.md"),
-            `# This table contains data about`
-          );
-          fs.writeFileSync(
-            path.join(projectDir, "definitions/table.sqlx"),
-            `config {
-            type: "table",
-            description: getContents('./descriptions.md'),
-          }
-            SELECT 1 AS test`
-          );
-
-          const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-          expect(
-            result.compile.compiledGraph.tables[0].actionDescriptor.description
-          ).to.equal(`# This table contains data about`);
-
-      });
-
-       test("throws error for invalid missing markedown", () => {
-          const projectDir = tmpDirFixture.createNewTmpDir();
-          fs.writeFileSync(
-            path.join(projectDir, "workflow_settings.yaml"),
-            VALID_WORKFLOW_SETTINGS_YAML
-          );
-          fs.mkdirSync(path.join(projectDir, "definitions"));
-          fs.writeFileSync(
-            path.join(projectDir, "definitions/table.sqlx"),
-            `config {
-            type: "table",
-            description: getContents('./nonexistent.md'),
-          }
-            SELECT 1 AS test`
-          );
-
-          const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-          expect(
-            result.compile.compiledGraph.graphErrors.compilationErrors[0].message
-          ).to.include("nonexistent.md");
-
-      });
-
-      test("throws error for file outisde of rootDir", () => {
-          const projectDir = tmpDirFixture.createNewTmpDir();
-          fs.writeFileSync(
-            path.join(projectDir, "workflow_settings.yaml"),
-            VALID_WORKFLOW_SETTINGS_YAML
-          );
-          fs.mkdirSync(path.join(projectDir, "definitions"));
-          fs.writeFileSync(
-            path.join(projectDir, "definitions/table.sqlx"),
-            `config {
-            type: "table",
-            description: getContents('../../description.md'),
-          }
-            SELECT 1 AS test`
-          );
-
-          const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-          expect(
-            result.compile.compiledGraph.graphErrors.compilationErrors[0].message
-          ).to.include("outside the project directory");
-
-      });
-    });
-    suite("invalid options", () => {
-      [
-        {
-          testName: "partitionBy invalid for BigQuery non-materialized views",
-          fileContents: `
-publish("name", {
-  type: "view",
-  materialized: false,
-  bigquery: {
-    partitionBy: "some_partition"
-  }
-})`,
-          expectedError:
-            'partitionBy/clusterBy can be applied only to materialized views'
-        },
-        {
-          testName: "clusterBy invalid for BigQuery non-materialized views",
-          fileContents: `
-publish("name", {
-  type: "view",
-  materialized: false,
-  bigquery: {
-    clusterBy: ["some_cluster"]
-  }
-})`,
-          expectedError:
-            'partitionBy/clusterBy can be applied only to materialized views'
-        },
-        {
-          testName: "partitionExpirationDays invalid for BigQuery views",
-          fileContents: `
-publish("name", {
-  type: "view",
-  bigquery: {
-    partitionExpirationDays: 7
-  }
-})`,
-          expectedError:
-            'Unexpected property "partitionExpirationDays" in BigQuery view config. Supported properties are: ["labels","additionalOptions","partitionBy","clusterBy"]'
-        },
-        {
-          testName: "requirePartitionFilter invalid for BigQuery views",
-          fileContents: `
-publish("name", {
-  type: "view",
-  bigquery: {
-    requirePartitionFilter: true
-  }
-})`,
-          expectedError:
-            'Unexpected property "requirePartitionFilter" in BigQuery view config. Supported properties are: ["labels","additionalOptions","partitionBy","clusterBy"]'
-        },
-        {
-          testName: "partitionExpirationDays invalid for BigQuery materialized views",
-          fileContents: `
-publish("name", {
-  type: "view",
-  materialized: true,
-  bigquery: {
-    partitionExpirationDays: 7
-  }
-})`,
-          expectedError:
-            'Unexpected property "partitionExpirationDays" in BigQuery view config. Supported properties are: ["labels","additionalOptions","partitionBy","clusterBy"]'
-        },
-        {
-          testName: "requirePartitionFilter invalid for BigQuery materialized views",
-          fileContents: `
-publish("name", {
-  type: "view",
-  materialized: true,
-  bigquery: {
-    requirePartitionFilter: true
-  }
-})`,
-          expectedError:
-            'Unexpected property "requirePartitionFilter" in BigQuery view config. Supported properties are: ["labels","additionalOptions","partitionBy","clusterBy"]'
-        },
-        {
-          testName: "materialized invalid for BigQuery tables",
-          fileContents: `
-publish("name", {
-  type: "table",
-  materialized: true,
-})`,
-          expectedError:
-            'Unexpected property "materialized", or property value type of "boolean" is incorrect. See https://dataform-co.github.io/dataform/docs/configs-reference#dataform-ActionConfig-TableConfig for allowed properties.'
-        },
-        {
-          testName: "partitionExpirationDays invalid for BigQuery tables",
-          fileContents: `
-publish("name", {
-  type: "table",
-  bigquery: {
-    partitionExpirationDays: 7
-  }
-})`,
-          expectedError:
-            "requirePartitionFilter/partitionExpirationDays are not valid for non partitioned BigQuery tables"
-        },
-        {
-          testName: "duplicate partitionExpirationDays is invalid",
-          fileContents: `
-publish("name", {
-  type: "table",
-  bigquery: {
-    partitionBy: "partition",
-    partitionExpirationDays: 1,
-    additionalOptions: {
-      partition_expiration_days: "7"
-    }
-  }
-})`,
-          expectedError: "partitionExpirationDays has been declared twice"
-        },
-        {
-          testName: "duplicate requirePartitionFilter is invalid",
-          fileContents: `
-publish("name", {
-  type: "table",
-  bigquery: {
-    partitionBy: "partition",
-    requirePartitionFilter: true,
-    additionalOptions: {
-      require_partition_filter: "false"
-    }
-  }
-})`,
-          expectedError: "requirePartitionFilter has been declared twice"
-        }
-      ].forEach(testParameters => {
-        test(testParameters.testName, () => {
-          const projectDir = tmpDirFixture.createNewTmpDir();
-          fs.writeFileSync(
-            path.join(projectDir, "workflow_settings.yaml"),
-            VALID_WORKFLOW_SETTINGS_YAML
-          );
-          fs.mkdirSync(path.join(projectDir, "definitions"));
-          fs.writeFileSync(path.join(projectDir, "definitions/operation.sqlx"), "SELECT 1");
-          fs.writeFileSync(
-            path.join(projectDir, `definitions/file.js`),
-            testParameters.fileContents
-          );
-
-          const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-          expect(
-            result.compile.compiledGraph.graphErrors.compilationErrors.map(
-              compilationError => compilationError.message
-            )
-          ).deep.equals([testParameters.expectedError]);
-        });
-      });
-    });
-
-    suite(`legacy publish().type() can still be called`, () => {
-      ["table", "incremental", "view"].forEach(fromType => {
-        ["table", "incremental", "view"].forEach(toType => {
-          test(`from type ${fromType} to ${toType}`, () => {
-            const projectDir = tmpDirFixture.createNewTmpDir();
-            fs.writeFileSync(
-              path.join(projectDir, "workflow_settings.yaml"),
-              dumpYaml(
-                dataform.WorkflowSettings.create(
-                  WorkflowSettingsTemplates.bigqueryWithDefaultProjectAndDataset
-                )
-              )
-            );
-            fs.mkdirSync(path.join(projectDir, "definitions"));
-            fs.writeFileSync(
-              path.join(projectDir, "definitions/publish.js"),
-              `
-publish("name", {type: "${fromType}", schema: "schemaOverride"}).type("${toType}")`
-            );
-
-            const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-            expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-            const tables = asPlainObject(result.compile.compiledGraph.tables);
-            expect(tables.length).equals(1);
-            expect(tables[0].type).equals(toType);
-            expect(tables[0].enumType).equals(toType.toUpperCase());
-
-            // Config options are carried over to the new table type, where possible.
-            expect(tables[0].target.schema).equals("schemaOverride");
-          });
-        });
-      });
-    });
-  });
-
-  suite("extensions interface", () => {
-    function setUpProjectWithExtension() {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        dumpYaml(dataform.WorkflowSettings.create(WorkflowSettingsTemplates.bigquery))
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(path.join(projectDir, "definitions/e.sqlx"), `config {type: "view"}`);
-      fs.writeFileSync(path.join(projectDir, "definitions/file.sqlx"), "${resolve('e')}");
-
-      return projectDir;
-    }
-
-    test("keeps regular compilation flow if unspecified", () => {
-      const projectDir = setUpProjectWithExtension();
-      const request = coreExecutionRequestFromPath(projectDir);
-
-      const result = runMainInVm(request);
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(result.compile.compiledGraph.targets?.map(t => t.name)).deep.equals(["e", "file"]);
-    });
-
-    test("keeps regular compilation flow if extension is disabled", () => {
-      const projectDir = setUpProjectWithExtension();
-      const request = coreExecutionRequestFromPath(projectDir);
-      request.compile.compileConfig.extension = {
-        name: "some-extension",
-        compilationMode: dataform.ExtensionCompilationMode.COMPILATION_MODE_UNSPECIFIED,
-      };
-
-      const result = runMainInVm(request);
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(result.compile.compiledGraph.targets?.map(t => t.name)).deep.equals(["e", "file"]);
-    });
-
-    test("runs prologue before regular compilation", () => {
-      const projectDir = setUpProjectWithExtension();
-      const request = coreExecutionRequestFromPath(projectDir);
-      request.compile.compileConfig.extension = {
-        name: "@dataform/sample-extension",
-        compilationMode: dataform.ExtensionCompilationMode.PROLOGUE,
-      };
-
-      const result = runMainInVm(request);
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(result.compile.compiledGraph.targets?.map(t => t.name)).deep.equals(["sample-action", "e", "file"]);
-    });
-
-    test("replaces regular compilation in application code mode", () => {
-      const projectDir = setUpProjectWithExtension();
-      const request = coreExecutionRequestFromPath(projectDir);
-      request.compile.compileConfig.extension = {
-        name: "@dataform/sample-extension",
-        compilationMode: dataform.ExtensionCompilationMode.APPLICATION_CODE,
-      };
-
-      const result = runMainInVm(request);
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(result.compile.compiledGraph.targets?.map(t => t.name)).deep.equals(["sample-action"]);
-    });
-
-    test("works in application mode without workflow settings", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(path.join(projectDir, "definitions/e.sqlx"), `config {type: "view"}`);
-      fs.writeFileSync(path.join(projectDir, "definitions/file.sqlx"), "${resolve('e')}");
-
-      const request = coreExecutionRequestFromPath(projectDir);
-      request.compile.compileConfig.extension = {
-        name: "@dataform/sample-extension",
-        compilationMode: dataform.ExtensionCompilationMode.APPLICATION_CODE
-      };
-
-      const result = runMainInVm(request);
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(result.compile.compiledGraph.targets?.map(t => t.name)).deep.equals([
-        "sample-action"
-      ]);
-    });
-
-    test("works in prologue mode without workflow settings", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(path.join(projectDir, "definitions/e.sqlx"), `config {type: "view"}`);
-      fs.writeFileSync(path.join(projectDir, "definitions/file.sqlx"), "${resolve('e')}");
-
-      const request = coreExecutionRequestFromPath(projectDir);
-      request.compile.compileConfig.extension = {
-        name: "@dataform/sample-extension",
-        compilationMode: dataform.ExtensionCompilationMode.PROLOGUE,
-      };
-
-      const result = runMainInVm(request);
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(result.compile.compiledGraph.targets?.map(t => t.name)).deep.equals(["sample-action", "e", "file"]);
-    });
-
-    test("catches extension import exceptions", () => {
-      const projectDir = setUpProjectWithExtension();
-      const request = coreExecutionRequestFromPath(projectDir);
-      request.compile.compileConfig.extension = {
-        name: "does-not-exist",
-        compilationMode: dataform.ExtensionCompilationMode.PROLOGUE,
-      };
-
-      const result = runMainInVm(request);
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors.length).equals(1);
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors[0].message).contains("Cannot find module");
-      expect(result.compile.compiledGraph.targets?.map(t => t.name)).deep.equals(["e", "file"]);
-    });
-
-    test("catches exceptions thrown from extension", () => {
-      const projectDir = setUpProjectWithExtension();
-      const request = coreExecutionRequestFromPath(projectDir, dataform.ProjectConfig.create({vars: {"throw-error": "true"}}));
-      request.compile.compileConfig.extension = {
-        name: "@dataform/sample-extension",
-        compilationMode: dataform.ExtensionCompilationMode.PROLOGUE,
-      };
-
-      const result = runMainInVm(request);
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors.length).equals(1);
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors[0].message).contains("throwing exception as requested!");
-      expect(result.compile.compiledGraph.targets?.map(t => t.name)).deep.equals(["e", "file"]);
-    });
-
-    test("persists extension compilation errors", () => {
-      const projectDir = setUpProjectWithExtension();
-      const request = coreExecutionRequestFromPath(projectDir, dataform.ProjectConfig.create({vars: {"store-compile-error": "true"}}));
-      request.compile.compileConfig.extension = {
-        name: "@dataform/sample-extension",
-        compilationMode: dataform.ExtensionCompilationMode.PROLOGUE,
-      };
-
-      const result = runMainInVm(request);
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors.length).equals(1);
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors[0].message).contains("storing compilation error as requested!");
-      expect(result.compile.compiledGraph.targets?.map(t => t.name)).deep.equals(["sample-action", "e", "file"]);
-    });
-    
-    test("preserveGovernanceControls propagates to compiled graph", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/file.sqlx"),
-        `
-config {
-  type: "table",
-  preserveGovernanceControls: true
-}
-select 1 as a`
-      );
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(result.compile.compiledGraph.tables[0].bigquery.preserveGovernanceControls).equals(true);
-    });
-
-    test("preserveGovernanceControls in workflow_settings.yaml propagates to compiled graph", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        `
-defaultProject: "project"
-defaultDataset: "dataset"
-preserveGovernanceControls: true`
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/file.sqlx"),
-        `
-config {
-  type: "table"
-}
-select 1 as a`
-      );
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(result.compile.compiledGraph.tables[0].bigquery.preserveGovernanceControls).equals(true);
-    });
-
-    test("preserveGovernanceControls on table overrides workflow_settings.yaml", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        `
-defaultProject: "project"
-defaultDataset: "dataset"
-preserveGovernanceControls: true`
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/file.sqlx"),
-        `
-config {
-  type: "table",
-  partitionBy: "somePartition",
-  preserveGovernanceControls: false
-}
-select 1 as a`
-      );
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(result.compile.compiledGraph.tables[0].bigquery.preserveGovernanceControls).equals(false);
-    });
-  });
-
-  suite("property graphs", () => {
-    const graphProjectConfig = {
-      warehouse: "bigquery",
-      defaultSchema: "defaultDataset",
-      defaultDatabase: "defaultProject",
-      defaultLocation: "US"
-    };
-    const graphStackTail = "\n    at CallSite {}".repeat(10);
-    const graphError = (fileName: string, message: string, extra: object = {}) => ({
-      fileName,
-      message,
-      stack: `Error: ${message}${graphStackTail}`,
-      ...extra
-    });
-
-    test("valid graph.yaml compiles end-to-end", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/graph.yaml"),
-        `
-name: SimpleGraph
-entities:
-- name: Customer
-  dataSourceString: defaultProject.defaultDataset.customers
-  keys:
-  - id
-`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(asPlainObject(result.compile.compiledGraph)).deep.equals(asPlainObject({
-        projectConfig: graphProjectConfig,
-        graphErrors: {},
-        dataformCoreVersion: version,
-        targets: [
-          { schema: "defaultDataset", name: "SimpleGraph", database: "defaultProject" }
-        ],
-        jitData: {},
-        propertyGraphs: [
-          {
-            target: {
-              schema: "defaultDataset",
-              name: "SimpleGraph",
-              database: "defaultProject"
-            },
-            canonicalTarget: {
-              schema: "defaultDataset",
-              name: "SimpleGraph",
-              database: "defaultProject"
-            },
-            fileName: "definitions/graph.yaml",
-            description: "",
-            disabled: false,
-            entities: [
-              {
-                name: "Customer",
-                dataSource: {
-                  schema: "defaultDataset",
-                  name: "customers",
-                  database: "defaultProject"
-                },
-                keys: ["id"]
-              }
-            ],
-            graphBody:
-              "NODE TABLES (\n" +
-              "  `defaultProject.defaultDataset.customers` AS Customer KEY (id)\n" +
-              ")"
-          }
-        ]
-      }));
-    });
-
-    test("graph.yaml tags propagate into the compiled proto", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/graph.yaml"),
-        `
-name: TaggedGraph
-tags:
-- nightly
-- reporting
-entities:
-- name: Customer
-  dataSourceString: defaultProject.defaultDataset.customers
-  keys:
-  - id
-`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(asPlainObject(result.compile.compiledGraph)).deep.equals(asPlainObject({
-        projectConfig: graphProjectConfig,
-        graphErrors: {},
-        dataformCoreVersion: version,
-        targets: [
-          { schema: "defaultDataset", name: "TaggedGraph", database: "defaultProject" }
-        ],
-        jitData: {},
-        propertyGraphs: [
-          {
-            target: {
-              schema: "defaultDataset",
-              name: "TaggedGraph",
-              database: "defaultProject"
-            },
-            canonicalTarget: {
-              schema: "defaultDataset",
-              name: "TaggedGraph",
-              database: "defaultProject"
-            },
-            fileName: "definitions/graph.yaml",
-            description: "",
-            disabled: false,
-            tags: ["nightly", "reporting"],
-            entities: [
-              {
-                name: "Customer",
-                dataSource: {
-                  schema: "defaultDataset",
-                  name: "customers",
-                  database: "defaultProject"
-                },
-                keys: ["id"]
-              }
-            ],
-            graphBody:
-              "NODE TABLES (\n" +
-              "  `defaultProject.defaultDataset.customers` AS Customer KEY (id)\n" +
-              ")"
-          }
-        ]
-      }));
-    });
-
-    test("more than one graph.yaml is rejected", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.mkdirSync(path.join(projectDir, "definitions/subdir"));
-      const graphBody = `
-name: MultiGraph
-entities:
-- name: Node
-  dataSourceString: defaultProject.defaultDataset.t
-  keys:
-  - id
-`;
-      fs.writeFileSync(path.join(projectDir, "definitions/graph.yaml"), graphBody);
-      fs.writeFileSync(path.join(projectDir, "definitions/subdir/graph.yaml"), graphBody);
-
-      const request = dataform.CoreExecutionRequest.create({
-        compile: {
-          compileConfig: {
-            projectDir: fs.realpathSync(projectDir),
-            filePaths: [
-              "workflow_settings.yaml",
-              "definitions/graph.yaml",
-              "definitions/subdir/graph.yaml"
-            ]
-          }
-        }
-      });
-
-      const result = runMainInVm(request);
-
-      expect(asPlainObject(result.compile.compiledGraph)).deep.equals(asPlainObject({
-        projectConfig: graphProjectConfig,
-        graphErrors: {
-          compilationErrors: [
-            graphError(
-              "definitions/graph.yaml",
-              "At most one graph.yaml is allowed per project (found 2: " +
-                "definitions/graph.yaml, definitions/subdir/graph.yaml). This " +
-                "restriction may be relaxed in a future version."
-            )
-          ]
-        },
-        dataformCoreVersion: version,
-        jitData: {}
-      }));
-    });
-
-    test("nodes-only graph compiles without EDGE TABLES", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/graph.yaml"),
-        `
-name: NodesOnly
-entities:
-- name: Customer
-  dataSourceString: defaultProject.defaultDataset.customers
-  keys:
-  - id
-- name: Product
-  dataSourceString: defaultProject.defaultDataset.products
-  keys:
-  - sku
-`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(asPlainObject(result.compile.compiledGraph)).deep.equals(asPlainObject({
-        projectConfig: graphProjectConfig,
-        graphErrors: {},
-        dataformCoreVersion: version,
-        targets: [
-          { schema: "defaultDataset", name: "NodesOnly", database: "defaultProject" }
-        ],
-        jitData: {},
-        propertyGraphs: [
-          {
-            target: {
-              schema: "defaultDataset",
-              name: "NodesOnly",
-              database: "defaultProject"
-            },
-            canonicalTarget: {
-              schema: "defaultDataset",
-              name: "NodesOnly",
-              database: "defaultProject"
-            },
-            fileName: "definitions/graph.yaml",
-            description: "",
-            disabled: false,
-            entities: [
-              {
-                name: "Customer",
-                dataSource: {
-                  schema: "defaultDataset",
-                  name: "customers",
-                  database: "defaultProject"
-                },
-                keys: ["id"]
-              },
-              {
-                name: "Product",
-                dataSource: {
-                  schema: "defaultDataset",
-                  name: "products",
-                  database: "defaultProject"
-                },
-                keys: ["sku"]
-              }
-            ],
-            graphBody:
-              "NODE TABLES (\n" +
-              "  `defaultProject.defaultDataset.customers` AS Customer KEY (id),\n" +
-              "  `defaultProject.defaultDataset.products` AS Product KEY (sku)\n" +
-              ")"
-          }
-        ]
-      }));
-    });
-
-    test("targetDataset overrides the schema on the graph target", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/graph.yaml"),
-        `
-name: CustomDsGraph
-targetDataset:
-  datasetId: customDs
-entities:
-- name: Customer
-  dataSourceString: defaultProject.defaultDataset.customers
-  keys:
-  - id
-`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(asPlainObject(result.compile.compiledGraph)).deep.equals(asPlainObject({
-        projectConfig: graphProjectConfig,
-        graphErrors: {},
-        dataformCoreVersion: version,
-        targets: [
-          { schema: "customDs", name: "CustomDsGraph", database: "defaultProject" }
-        ],
-        jitData: {},
-        propertyGraphs: [
-          {
-            target: {
-              schema: "customDs",
-              name: "CustomDsGraph",
-              database: "defaultProject"
-            },
-            canonicalTarget: {
-              schema: "customDs",
-              name: "CustomDsGraph",
-              database: "defaultProject"
-            },
-            fileName: "definitions/graph.yaml",
-            description: "",
-            disabled: false,
-            entities: [
-              {
-                name: "Customer",
-                dataSource: {
-                  schema: "defaultDataset",
-                  name: "customers",
-                  database: "defaultProject"
-                },
-                keys: ["id"]
-              }
-            ],
-            graphBody:
-              "NODE TABLES (\n" +
-              "  `defaultProject.defaultDataset.customers` AS Customer KEY (id)\n" +
-              ")"
-          }
-        ]
-      }));
-    });
-
-    test("empty graph.yaml produces a compilation error", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(path.join(projectDir, "definitions/graph.yaml"), "");
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(asPlainObject(result.compile.compiledGraph)).deep.equals(asPlainObject({
-        projectConfig: graphProjectConfig,
-        graphErrors: {
-          compilationErrors: [
-            graphError(
-              "definitions/graph.yaml",
-              "Property graph config is empty or malformed. Expected a top-level " +
-                "object with 'name' and 'entities'."
-            )
-          ]
-        },
-        dataformCoreVersion: version,
-        jitData: {}
-      }));
-    });
-
-    test("graph.yaml with only a comment produces a compilation error", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/graph.yaml"),
-        "# nothing here\n"
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(asPlainObject(result.compile.compiledGraph)).deep.equals(asPlainObject({
-        projectConfig: graphProjectConfig,
-        graphErrors: {
-          compilationErrors: [
-            graphError(
-              "definitions/graph.yaml",
-              "Property graph config is empty or malformed. Expected a top-level " +
-                "object with 'name' and 'entities'."
-            )
-          ]
-        },
-        dataformCoreVersion: version,
-        jitData: {}
-      }));
-    });
-
-    test("graph.yaml with a top-level scalar produces a compilation error", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/graph.yaml"),
-        "just a string\n"
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(asPlainObject(result.compile.compiledGraph)).deep.equals(asPlainObject({
-        projectConfig: graphProjectConfig,
-        graphErrors: {
-          compilationErrors: [
-            graphError(
-              "definitions/graph.yaml",
-              "Property graph config is empty or malformed. Expected a top-level " +
-                "object with 'name' and 'entities'."
-            )
-          ]
-        },
-        dataformCoreVersion: version,
-        jitData: {}
-      }));
-    });
-
-    test("graph.yaml missing entities produces a compilation error", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/graph.yaml"),
-        `
-name: EmptyGraph
-`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(asPlainObject(result.compile.compiledGraph)).deep.equals(asPlainObject({
-        projectConfig: graphProjectConfig,
-        graphErrors: {
-          compilationErrors: [
-            graphError(
-              "definitions/graph.yaml",
-              "Property graph 'EmptyGraph' must declare at least one entity."
-            )
-          ]
-        },
-        dataformCoreVersion: version,
-        jitData: {}
-      }));
-    });
-
-    test("graph with relationships emits EDGE TABLES", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/graph.yaml"),
-        `
-name: RelGraph
-entities:
-- name: Customer
-  dataSourceString: defaultProject.defaultDataset.customers
-  keys:
-  - id
-- name: Order
-  dataSourceString: defaultProject.defaultDataset.orders
-  keys:
-  - id
-relationships:
-- name: PlacedBy
-  dataSourceString: defaultProject.defaultDataset.orders
-  source:
-    entity: Order
-    joinKeys:
-    - order_id
-  destination:
-    entity: Customer
-    joinKeys:
-    - customer_id
-`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(asPlainObject(result.compile.compiledGraph)).deep.equals(asPlainObject({
-        projectConfig: graphProjectConfig,
-        graphErrors: {},
-        dataformCoreVersion: version,
-        targets: [
-          { schema: "defaultDataset", name: "RelGraph", database: "defaultProject" }
-        ],
-        jitData: {},
-        propertyGraphs: [
-          {
-            target: {
-              schema: "defaultDataset",
-              name: "RelGraph",
-              database: "defaultProject"
-            },
-            canonicalTarget: {
-              schema: "defaultDataset",
-              name: "RelGraph",
-              database: "defaultProject"
-            },
-            fileName: "definitions/graph.yaml",
-            description: "",
-            disabled: false,
-            entities: [
-              {
-                name: "Customer",
-                dataSource: {
-                  schema: "defaultDataset",
-                  name: "customers",
-                  database: "defaultProject"
-                },
-                keys: ["id"]
-              },
-              {
-                name: "Order",
-                dataSource: {
-                  schema: "defaultDataset",
-                  name: "orders",
-                  database: "defaultProject"
-                },
-                keys: ["id"]
-              }
-            ],
-            relationships: [
-              {
-                name: "PlacedBy",
-                dataSource: {
-                  schema: "defaultDataset",
-                  name: "orders",
-                  database: "defaultProject"
-                },
-                source: {
-                  entity: "Order",
-                  relationshipColumns: ["order_id"],
-                  entityColumns: ["id"]
-                },
-                destination: {
-                  entity: "Customer",
-                  relationshipColumns: ["customer_id"],
-                  entityColumns: ["id"]
-                }
-              }
-            ],
-            graphBody:
-              "NODE TABLES (\n" +
-              "  `defaultProject.defaultDataset.customers` AS Customer KEY (id),\n" +
-              "  `defaultProject.defaultDataset.orders` AS Order KEY (id)\n" +
-              ")\n" +
-              "EDGE TABLES (\n" +
-              "  `defaultProject.defaultDataset.orders` AS PlacedBy " +
-              "SOURCE KEY (order_id) REFERENCES Order (id) " +
-              "DESTINATION KEY (customer_id) REFERENCES Customer (id)\n" +
-              ")"
-          }
-        ]
-      }));
-    });
-
-    test("ref to declaration resolves entity dataSource and renders graphBody", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- declaration:
-    name: books
-`
-      );
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/graph.yaml"),
-        `
-name: RefGraph
-entities:
-- name: Book
-  ref: books
-  keys:
-  - id
-`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.propertyGraphs)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              schema: "defaultDataset",
-              name: "RefGraph",
-              database: "defaultProject"
-            },
-            canonicalTarget: {
-              schema: "defaultDataset",
-              name: "RefGraph",
-              database: "defaultProject"
-            },
-            dependencyTargets: [
-              {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "books"
-              }
-            ],
-            fileName: "definitions/graph.yaml",
-            description: "",
-            disabled: false,
-            entities: [
-              {
-                name: "Book",
-                dataSource: {
-                  schema: "defaultDataset",
-                  name: "books",
-                  database: "defaultProject"
-                },
-                keys: ["id"]
-              }
-            ],
-            graphBody:
-              "NODE TABLES (\n" +
-              "  `defaultProject.defaultDataset.books` AS Book KEY (id)\n" +
-              ")"
-          }
-        ])
-      );
-    });
-
-    test("ref with schema override resolves the matching declaration", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- declaration:
-    name: books
-    dataset: alt
-- declaration:
-    name: books
-`
-      );
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/graph.yaml"),
-        `
-name: RefWithSchemaGraph
-entities:
-- name: Book
-  ref:
-    name: books
-    schema: alt
-  keys:
-  - id
-`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.propertyGraphs)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              schema: "defaultDataset",
-              name: "RefWithSchemaGraph",
-              database: "defaultProject"
-            },
-            canonicalTarget: {
-              schema: "defaultDataset",
-              name: "RefWithSchemaGraph",
-              database: "defaultProject"
-            },
-            dependencyTargets: [
-              {
-                database: "defaultProject",
-                schema: "alt",
-                name: "books"
-              }
-            ],
-            fileName: "definitions/graph.yaml",
-            description: "",
-            disabled: false,
-            entities: [
-              {
-                name: "Book",
-                dataSource: {
-                  schema: "alt",
-                  name: "books",
-                  database: "defaultProject"
-                },
-                keys: ["id"]
-              }
-            ],
-            graphBody:
-              "NODE TABLES (\n" +
-              "  `defaultProject.alt.books` AS Book KEY (id)\n" +
-              ")"
-          }
-        ])
-      );
-    });
-
-    test("ref with includeDependentAssertions pulls the dependency's assertions", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/books.sqlx"),
-        `config {
-  type: "table",
-  assertions: { rowConditions: ["id > 0"] }
-}
-select 1 as id`
-      );
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/graph.yaml"),
-        `
-name: AssertRefGraph
-entities:
-- name: Book
-  ref:
-    name: books
-    includeDependentAssertions: true
-  keys:
-  - id
-`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.propertyGraphs)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              schema: "defaultDataset",
-              name: "AssertRefGraph",
-              database: "defaultProject"
-            },
-            canonicalTarget: {
-              schema: "defaultDataset",
-              name: "AssertRefGraph",
-              database: "defaultProject"
-            },
-            dependencyTargets: [
-              {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "books"
-              },
-              {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "defaultDataset_books_assertions_rowConditions"
-              }
-            ],
-            fileName: "definitions/graph.yaml",
-            description: "",
-            disabled: false,
-            entities: [
-              {
-                name: "Book",
-                dataSource: {
-                  schema: "defaultDataset",
-                  name: "books",
-                  database: "defaultProject"
-                },
-                keys: ["id"]
-              }
-            ],
-            graphBody:
-              "NODE TABLES (\n" +
-              "  `defaultProject.defaultDataset.books` AS Book KEY (id)\n" +
-              ")"
-          }
-        ])
-      );
-    });
-
-    test("graph-level dependOnDependencyAssertions pulls every ref's assertions", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/books.sqlx"),
-        `config {
-  type: "table",
-  assertions: { rowConditions: ["id > 0"] }
-}
-select 1 as id`
-      );
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/graph.yaml"),
-        `
-name: GraphAssertDefaultGraph
-dependOnDependencyAssertions: true
-entities:
-- name: Book
-  ref: books
-  keys:
-  - id
-`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.propertyGraphs)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              schema: "defaultDataset",
-              name: "GraphAssertDefaultGraph",
-              database: "defaultProject"
-            },
-            canonicalTarget: {
-              schema: "defaultDataset",
-              name: "GraphAssertDefaultGraph",
-              database: "defaultProject"
-            },
-            dependencyTargets: [
-              {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "books"
-              },
-              {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "defaultDataset_books_assertions_rowConditions"
-              }
-            ],
-            fileName: "definitions/graph.yaml",
-            description: "",
-            disabled: false,
-            entities: [
-              {
-                name: "Book",
-                dataSource: {
-                  schema: "defaultDataset",
-                  name: "books",
-                  database: "defaultProject"
-                },
-                keys: ["id"]
-              }
-            ],
-            graphBody:
-              "NODE TABLES (\n" +
-              "  `defaultProject.defaultDataset.books` AS Book KEY (id)\n" +
-              ")"
-          }
-        ])
-      );
-    });
-
-    test("missing ref emits a compilation error and leaves graphBody empty", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/graph.yaml"),
-        `
-name: MissingRefGraph
-entities:
-- name: Book
-  ref: nonexistent
-  keys:
-  - id
-`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      const missingRefTarget = {
-        schema: "defaultDataset",
-        name: "MissingRefGraph",
-        database: "defaultProject"
-      };
-      expect(asPlainObject(result.compile.compiledGraph)).deep.equals(asPlainObject({
-        projectConfig: graphProjectConfig,
-        graphErrors: {
-          compilationErrors: [
-            graphError(
-              "definitions/graph.yaml",
-              "Missing dependency detected: Action " +
-                "\"defaultProject.defaultDataset.MissingRefGraph\" depends on " +
-                "\"{\"name\":\"nonexistent\",\"includeDependentAssertions\":false}\" " +
-                "which does not exist",
-              {
-                actionName: "defaultProject.defaultDataset.MissingRefGraph",
-                actionTarget: missingRefTarget
-              }
-            )
-          ]
-        },
-        dataformCoreVersion: version,
-        targets: [missingRefTarget],
-        jitData: {},
-        propertyGraphs: [
-          {
-            target: missingRefTarget,
-            canonicalTarget: missingRefTarget,
-            fileName: "definitions/graph.yaml",
-            description: "",
-            disabled: false,
-            entities: [
-              {
-                name: "Book",
-                keys: ["id"]
-              }
-            ]
-          }
-        ]
-      }));
-    });
-
-    test("ref to a table respects datasetSuffix on the resolved dependency", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        `
-defaultProject: defaultProject
-defaultDataset: defaultDataset
-defaultLocation: US
-datasetSuffix: dev
-`
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/books.sqlx"),
-        `config {type: "table"}
-select 1 as id`
-      );
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/graph.yaml"),
-        `
-name: SuffixRefGraph
-entities:
-- name: Book
-  ref: books
-  keys:
-  - id
-`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.propertyGraphs)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              schema: "defaultDataset_dev",
-              name: "SuffixRefGraph",
-              database: "defaultProject"
-            },
-            canonicalTarget: {
-              schema: "defaultDataset",
-              name: "SuffixRefGraph",
-              database: "defaultProject"
-            },
-            dependencyTargets: [
-              {
-                schema: "defaultDataset_dev",
-                name: "books",
-                database: "defaultProject"
-              }
-            ],
-            fileName: "definitions/graph.yaml",
-            description: "",
-            disabled: false,
-            entities: [
-              {
-                name: "Book",
-                dataSource: {
-                  schema: "defaultDataset_dev",
-                  name: "books",
-                  database: "defaultProject"
-                },
-                keys: ["id"]
-              }
-            ],
-            graphBody:
-              "NODE TABLES (\n" +
-              "  `defaultProject.defaultDataset_dev.books` AS Book KEY (id)\n" +
-              ")"
-          }
-        ])
-      );
-    });
-
-    test("ref with database override resolves the matching declaration", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- declaration:
-    name: books
-    project: otherProject
-- declaration:
-    name: books
-`
-      );
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/graph.yaml"),
-        `
-name: RefWithDatabaseGraph
-entities:
-- name: Book
-  ref:
-    name: books
-    database: otherProject
-  keys:
-  - id
-`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.propertyGraphs)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              schema: "defaultDataset",
-              name: "RefWithDatabaseGraph",
-              database: "defaultProject"
-            },
-            canonicalTarget: {
-              schema: "defaultDataset",
-              name: "RefWithDatabaseGraph",
-              database: "defaultProject"
-            },
-            dependencyTargets: [
-              {
-                database: "otherProject",
-                schema: "defaultDataset",
-                name: "books"
-              }
-            ],
-            fileName: "definitions/graph.yaml",
-            description: "",
-            disabled: false,
-            entities: [
-              {
-                name: "Book",
-                dataSource: {
-                  database: "otherProject",
-                  schema: "defaultDataset",
-                  name: "books"
-                },
-                keys: ["id"]
-              }
-            ],
-            graphBody:
-              "NODE TABLES (\n" +
-              "  `otherProject.defaultDataset.books` AS Book KEY (id)\n" +
-              ")"
-          }
-        ])
-      );
-    });
-
-    test("relationship ref resolves through the full pipeline", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- declaration:
-    name: wrote
-`
-      );
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/graph.yaml"),
-        `
-name: RelationshipRefGraph
-entities:
-- name: Book
-  dataSourceString: defaultProject.defaultDataset.books
-  keys:
-  - id
-- name: Author
-  dataSourceString: defaultProject.defaultDataset.authors
-  keys:
-  - id
-relationships:
-- name: WrittenBy
-  ref: wrote
-  keys:
-  - author_id
-  - book_id
-  source:
-    entity: Book
-    joinKeys:
-    - book_id
-  destination:
-    entity: Author
-    joinKeys:
-    - author_id
-`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.propertyGraphs)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              schema: "defaultDataset",
-              name: "RelationshipRefGraph",
-              database: "defaultProject"
-            },
-            canonicalTarget: {
-              schema: "defaultDataset",
-              name: "RelationshipRefGraph",
-              database: "defaultProject"
-            },
-            dependencyTargets: [
-              {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "wrote"
-              }
-            ],
-            fileName: "definitions/graph.yaml",
-            description: "",
-            disabled: false,
-            entities: [
-              {
-                name: "Book",
-                dataSource: {
-                  database: "defaultProject",
-                  schema: "defaultDataset",
-                  name: "books"
-                },
-                keys: ["id"]
-              },
-              {
-                name: "Author",
-                dataSource: {
-                  database: "defaultProject",
-                  schema: "defaultDataset",
-                  name: "authors"
-                },
-                keys: ["id"]
-              }
-            ],
-            relationships: [
-              {
-                name: "WrittenBy",
-                dataSource: {
-                  database: "defaultProject",
-                  schema: "defaultDataset",
-                  name: "wrote"
-                },
-                keys: ["author_id", "book_id"],
-                source: {
-                  entity: "Book",
-                  relationshipColumns: ["book_id"],
-                  entityColumns: ["id"]
-                },
-                destination: {
-                  entity: "Author",
-                  relationshipColumns: ["author_id"],
-                  entityColumns: ["id"]
-                }
-              }
-            ],
-            graphBody:
-              "NODE TABLES (\n" +
-              "  `defaultProject.defaultDataset.books` AS Book KEY (id),\n" +
-              "  `defaultProject.defaultDataset.authors` AS Author KEY (id)\n" +
-              ")\n" +
-              "EDGE TABLES (\n" +
-              "  `defaultProject.defaultDataset.wrote` AS WrittenBy " +
-              "KEY (author_id, book_id) " +
-              "SOURCE KEY (book_id) REFERENCES Book (id) " +
-              "DESTINATION KEY (author_id) REFERENCES Author (id)\n" +
-              ")"
-          }
-        ])
-      );
-    });
-
-    test("ref to a view resolves and picks up datasetSuffix", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        `
-defaultProject: defaultProject
-defaultDataset: defaultDataset
-defaultLocation: US
-datasetSuffix: dev
-`
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/books.sqlx"),
-        `config {type: "view"}
-select 1 as id`
-      );
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/graph.yaml"),
-        `
-name: ViewRefGraph
-entities:
-- name: Book
-  ref: books
-  keys:
-  - id
-`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.propertyGraphs)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              schema: "defaultDataset_dev",
-              name: "ViewRefGraph",
-              database: "defaultProject"
-            },
-            canonicalTarget: {
-              schema: "defaultDataset",
-              name: "ViewRefGraph",
-              database: "defaultProject"
-            },
-            dependencyTargets: [
-              {
-                database: "defaultProject",
-                schema: "defaultDataset_dev",
-                name: "books"
-              }
-            ],
-            fileName: "definitions/graph.yaml",
-            description: "",
-            disabled: false,
-            entities: [
-              {
-                name: "Book",
-                dataSource: {
-                  database: "defaultProject",
-                  schema: "defaultDataset_dev",
-                  name: "books"
-                },
-                keys: ["id"]
-              }
-            ],
-            graphBody:
-              "NODE TABLES (\n" +
-              "  `defaultProject.defaultDataset_dev.books` AS Book KEY (id)\n" +
-              ")"
-          }
-        ])
-      );
-    });
-
-    test("ambiguous ref emits a compilation error", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- declaration:
-    name: books
-    dataset: one
-- declaration:
-    name: books
-    dataset: two
-`
-      );
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/graph.yaml"),
-        `
-name: AmbiguousRefGraph
-entities:
-- name: Book
-  ref: books
-  keys:
-  - id
-`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      const declOneTarget = { schema: "one", name: "books", database: "defaultProject" };
-      const declTwoTarget = { schema: "two", name: "books", database: "defaultProject" };
-      const graphTarget = {
-        schema: "defaultDataset",
-        name: "AmbiguousRefGraph",
-        database: "defaultProject"
-      };
-      expect(asPlainObject(result.compile.compiledGraph)).deep.equals(asPlainObject({
-        projectConfig: graphProjectConfig,
-        graphErrors: {
-          compilationErrors: [
-            graphError(
-              "definitions/graph.yaml",
-              `Ambiguous Action name: {"name":"books","includeDependentAssertions":false}. ` +
-                "Did you mean one of: one.books, two.books.",
-              {
-                actionName: "defaultProject.defaultDataset.AmbiguousRefGraph",
-                actionTarget: graphTarget
-              }
-            )
-          ]
-        },
-        dataformCoreVersion: version,
-        targets: [declOneTarget, declTwoTarget, graphTarget],
-        jitData: {},
-        declarations: [
-          { target: declOneTarget, canonicalTarget: declOneTarget },
-          { target: declTwoTarget, canonicalTarget: declTwoTarget }
-        ],
-        propertyGraphs: [
-          {
-            target: graphTarget,
-            canonicalTarget: graphTarget,
-            fileName: "definitions/graph.yaml",
-            description: "",
-            disabled: false,
-            entities: [{ name: "Book", keys: ["id"] }]
-          }
-        ]
-      }));
-    });
-
-    test("ref to a table respects projectSuffix on the resolved dependency", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        `
-defaultProject: defaultProject
-defaultDataset: defaultDataset
-defaultLocation: US
-projectSuffix: dev
-`
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/books.sqlx"),
-        `config {type: "table"}
-select 1 as id`
-      );
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/graph.yaml"),
-        `
-name: ProjectSuffixRefGraph
-entities:
-- name: Book
-  ref: books
-  keys:
-  - id
-`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.propertyGraphs)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              schema: "defaultDataset",
-              name: "ProjectSuffixRefGraph",
-              database: "defaultProject_dev"
-            },
-            canonicalTarget: {
-              schema: "defaultDataset",
-              name: "ProjectSuffixRefGraph",
-              database: "defaultProject"
-            },
-            dependencyTargets: [
-              {
-                schema: "defaultDataset",
-                name: "books",
-                database: "defaultProject_dev"
-              }
-            ],
-            fileName: "definitions/graph.yaml",
-            description: "",
-            disabled: false,
-            entities: [
-              {
-                name: "Book",
-                dataSource: {
-                  schema: "defaultDataset",
-                  name: "books",
-                  database: "defaultProject_dev"
-                },
-                keys: ["id"]
-              }
-            ],
-            graphBody:
-              "NODE TABLES (\n" +
-              "  `defaultProject_dev.defaultDataset.books` AS Book KEY (id)\n" +
-              ")"
-          }
-        ])
-      );
-    });
-
-    test("ref to a table respects namePrefix on the resolved dependency", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        `
-defaultProject: defaultProject
-defaultDataset: defaultDataset
-defaultLocation: US
-namePrefix: pfx
-`
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/books.sqlx"),
-        `config {type: "table"}
-select 1 as id`
-      );
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/graph.yaml"),
-        `
-name: NamePrefixRefGraph
-entities:
-- name: Book
-  ref: books
-  keys:
-  - id
-`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.propertyGraphs)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              schema: "defaultDataset",
-              name: "pfx_NamePrefixRefGraph",
-              database: "defaultProject"
-            },
-            canonicalTarget: {
-              schema: "defaultDataset",
-              name: "NamePrefixRefGraph",
-              database: "defaultProject"
-            },
-            dependencyTargets: [
-              {
-                schema: "defaultDataset",
-                name: "pfx_books",
-                database: "defaultProject"
-              }
-            ],
-            fileName: "definitions/graph.yaml",
-            description: "",
-            disabled: false,
-            entities: [
-              {
-                name: "Book",
-                dataSource: {
-                  schema: "defaultDataset",
-                  name: "pfx_books",
-                  database: "defaultProject"
-                },
-                keys: ["id"]
-              }
-            ],
-            graphBody:
-              "NODE TABLES (\n" +
-              "  `defaultProject.defaultDataset.pfx_books` AS Book KEY (id)\n" +
-              ")"
-          }
-        ])
-      );
-    });
-
-    test("graph target colliding with a table target is flagged as duplicate", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/collision.sqlx"),
-        `config {type: "table", name: "CollisionName"}
-select 1 as a`
-      );
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/graph.yaml"),
-        `
-name: CollisionName
-entities:
-- name: Customer
-  dataSourceString: defaultProject.defaultDataset.customers
-  keys:
-  - id
-`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      const collisionTarget = {
-        schema: "defaultDataset",
-        name: "CollisionName",
-        database: "defaultProject"
-      };
-      const collisionActionName = "defaultProject.defaultDataset.CollisionName";
-      const collisionTargetJson =
-        `{"schema":"defaultDataset","name":"CollisionName","database":"defaultProject"}`;
-      const duplicateActionMessage =
-        "Duplicate action name detected. Names within a schema must be unique " +
-        "across tables, declarations, assertions, and operations:\n" +
-        `"${collisionTargetJson}"`;
-      const duplicateCanonicalMessage =
-        "Duplicate canonical target detected. Canonical targets must be unique " +
-        "across tables, declarations, assertions, and operations:\n" +
-        `"${collisionTargetJson}"`;
-      expect(asPlainObject(result.compile.compiledGraph)).deep.equals(asPlainObject({
-        projectConfig: graphProjectConfig,
-        graphErrors: {
-          compilationErrors: [
-            graphError("definitions/collision.sqlx", duplicateActionMessage, {
-              actionName: collisionActionName,
-              actionTarget: collisionTarget
-            }),
-            graphError("definitions/collision.sqlx", duplicateCanonicalMessage, {
-              actionName: collisionActionName,
-              actionTarget: collisionTarget
-            }),
-            graphError("definitions/graph.yaml", duplicateActionMessage, {
-              actionName: collisionActionName,
-              actionTarget: collisionTarget
-            }),
-            graphError("definitions/graph.yaml", duplicateCanonicalMessage, {
-              actionName: collisionActionName,
-              actionTarget: collisionTarget
-            })
-          ]
-        },
-        dataformCoreVersion: version,
-        targets: [collisionTarget, collisionTarget],
-        jitData: {}
-      }));
-    });
-    test("graph.yaml accepts snake_case keys per BQ spec", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/graph.yaml"),
-        `
-name: SnakeGraph
-description: end to end snake case
-target_dataset:
-  project_id: defaultProject
-  dataset_id: defaultDataset
-entities:
-- name: Account
-  data_source_string: defaultProject.defaultDataset.accounts
-  keys:
-  - id
-  fields:
-    import_all: true
-    except:
-    - secret
-relationships:
-- name: Owns
-  data_source_string: defaultProject.defaultDataset.ownership
-  source:
-    entity: Account
-    join_keys:
-      relationship_columns:
-      - owner_id
-  destination:
-    entity: Account
-    join_keys:
-      relationship_columns:
-      - owned_id
-`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(asPlainObject(result.compile.compiledGraph)).deep.equals(asPlainObject({
-        projectConfig: graphProjectConfig,
-        graphErrors: {},
-        dataformCoreVersion: version,
-        targets: [
-          { schema: "defaultDataset", name: "SnakeGraph", database: "defaultProject" }
-        ],
-        jitData: {},
-        propertyGraphs: [
-          {
-            target: {
-              schema: "defaultDataset",
-              name: "SnakeGraph",
-              database: "defaultProject"
-            },
-            canonicalTarget: {
-              schema: "defaultDataset",
-              name: "SnakeGraph",
-              database: "defaultProject"
-            },
-            fileName: "definitions/graph.yaml",
-            description: "end to end snake case",
-            disabled: false,
-            entities: [
-              {
-                name: "Account",
-                dataSource: {
-                  schema: "defaultDataset",
-                  name: "accounts",
-                  database: "defaultProject"
-                },
-                keys: ["id"],
-                labels: [
-                  {
-                    name: "Account",
-                    description: "",
-                    importAll: true,
-                    importExcept: ["secret"],
-                    isDefault: true
-                  }
-                ]
-              }
-            ],
-            relationships: [
-              {
-                name: "Owns",
-                dataSource: {
-                  schema: "defaultDataset",
-                  name: "ownership",
-                  database: "defaultProject"
-                },
-                source: {
-                  entity: "Account",
-                  relationshipColumns: ["owner_id"],
-                  entityColumns: ["id"]
-                },
-                destination: {
-                  entity: "Account",
-                  relationshipColumns: ["owned_id"],
-                  entityColumns: ["id"]
-                }
-              }
-            ],
-            graphBody:
-              "NODE TABLES (\n" +
-              "  `defaultProject.defaultDataset.accounts` AS Account KEY (id) " +
-              "DEFAULT LABEL PROPERTIES ARE ALL COLUMNS EXCEPT (secret)\n" +
-              ")\n" +
-              "EDGE TABLES (\n" +
-              "  `defaultProject.defaultDataset.ownership` AS Owns " +
-              "SOURCE KEY (owner_id) REFERENCES Account (id) " +
-              "DESTINATION KEY (owned_id) REFERENCES Account (id)\n" +
-              ")"
-          }
-        ]
-      }));
-    });
-
-    test("mixed ref and dataSourceString: only ref target appears in dependencyTargets", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/books.sqlx"),
-        `config {type: "table"}
-select 1 as id`
-      );
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/authors.sqlx"),
-        `config {type: "table"}
-select 1 as id`
-      );
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/graph.yaml"),
-        `
-name: MixedRefStringGraph
-entities:
-- name: Book
-  ref: books
-  keys:
-  - id
-- name: Author
-  dataSourceString: defaultProject.defaultDataset.authors
-  keys:
-  - id
-`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.propertyGraphs)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              schema: "defaultDataset",
-              name: "MixedRefStringGraph",
-              database: "defaultProject"
-            },
-            canonicalTarget: {
-              schema: "defaultDataset",
-              name: "MixedRefStringGraph",
-              database: "defaultProject"
-            },
-            dependencyTargets: [
-              { database: "defaultProject", schema: "defaultDataset", name: "books" }
-            ],
-            fileName: "definitions/graph.yaml",
-            description: "",
-            disabled: false,
-            entities: [
-              {
-                name: "Book",
-                dataSource: {
-                  schema: "defaultDataset",
-                  name: "books",
-                  database: "defaultProject"
-                },
-                keys: ["id"]
-              },
-              {
-                name: "Author",
-                dataSource: {
-                  schema: "defaultDataset",
-                  name: "authors",
-                  database: "defaultProject"
-                },
-                keys: ["id"]
-              }
-            ],
-            graphBody:
-              "NODE TABLES (\n" +
-              "  `defaultProject.defaultDataset.books` AS Book KEY (id),\n" +
-              "  `defaultProject.defaultDataset.authors` AS Author KEY (id)\n" +
-              ")"
-          }
-        ])
-      );
-    });
-  });
-});
