+++ testing/BUILD
@@ -18,12 +18,17 @@ ts_library(
     ],
     deps = [
         "//:modules-fix",
+        "//protos:ts",
         "@npm//@types/diff",
+        "@npm//@types/fs-extra",
+        "@npm//@types/js-yaml",
         "@npm//@types/json-stable-stringify",
         "@npm//@types/node",
         "@npm//@types/rimraf",
         "@npm//chalk",
         "@npm//diff",
+        "@npm//fs-extra",
+        "@npm//js-yaml",
         "@npm//json-stable-stringify",
         "@npm//rimraf",
     ],
